% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-control.R
\name{accepted_controls_by_operation}
\alias{accepted_controls_by_operation}
\title{Return the Accepted Control Arguments by Operation}
\usage{
accepted_controls_by_operation(
  operation = c("create", "insert", "update", "upsert", "delete", "undelete",
    "hardDelete", "query", "queryall", "retrieve", "convertLead", "merge",
    "describeSObjects", "resetPassword")
)
}
\arguments{
\item{operation}{\code{character}; a string defining the type of operation being
performed (e.g. "insert", "update", "upsert", "delete", "hardDelete",
"query", "queryall").}
}
\value{
\code{character}; a vector of strings indicating which control arguments
are accepted by the specified operation.
}
\description{
Return the Accepted Control Arguments by Operation
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
