% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-xml.R
\name{build_manifest_xml_from_list}
\alias{build_manifest_xml_from_list}
\title{Bulk Binary Attachments Manifest List to XML Converter}
\usage{
build_manifest_xml_from_list(input_data, root = NULL)
}
\arguments{
\item{input_data}{\code{list}; data to be appended}

\item{root}{\code{xmlNode}; an XML node to be used as the root}
}
\value{
\code{xmlNode}; an XML node constructed into a manifest data required
by the Bulk APIs for handling binary attachment data.
}
\description{
This function converts a list of data for binary attachments to XML
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\references{
\url{https://developer.salesforce.com/docs/atlas.en-us.api_asynch.meta/api_asynch/binary_create_request_file.htm}
}
\keyword{internal}
