% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-xml.R
\name{build_metadata_xml_from_list}
\alias{build_metadata_xml_from_list}
\title{Metadata List to XML Converter}
\usage{
build_metadata_xml_from_list(
  input_data,
  metatype = NULL,
  root_name = NULL,
  ns = c(character(0)),
  root = NULL
)
}
\arguments{
\item{input_data}{XML document serving as the basis upon which to add the list}

\item{metatype}{a character indicating the element name of each record in the list}

\item{root_name}{character; the name of the root node if created}

\item{ns}{named vector; a collection of character strings indicating the namespace 
definitions of the root node if created}

\item{root}{\code{XMLNode}; a node to be used as the root}
}
\value{
A XML document with the sublist data added
}
\description{
This function converts a list of metadata to XML
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\concept{metadata salesforce api}
\keyword{internal}
