% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-httr.R
\name{build_proxy}
\alias{build_proxy}
\title{Function to build a proxy object to pass along with httr requests}
\usage{
build_proxy()
}
\description{
Function to build a proxy object to pass along with httr requests
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
