% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-httr.R
\name{catch_errors}
\alias{catch_errors}
\title{Function to catch and print HTTP errors}
\usage{
catch_errors(x)
}
\arguments{
\item{x}{\code{response()}; a response from an HTTP request}
}
\value{
\code{logical}; return \code{FALSE} if the function finishes without
detecting an error, otherwise stop the function call.
}
\description{
Function to catch and print HTTP errors
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
