% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-xml.R
\name{combine_parent_and_child_resultsets}
\alias{combine_parent_and_child_resultsets}
\title{Bind the records from nested parent-to-child queries}
\usage{
combine_parent_and_child_resultsets(parents_df, child_df_list)
}
\arguments{
\item{x}{\code{list}; a list of \code{xml_node} from an xml2 parsed response}
}
\description{
This function accepts a \code{data.frame} with one row representing each parent record 
returned by a query with a corresponding list element in the list of child 
record results stored as \code{tbl_df} in a list.
}
\keyword{internal}
