% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-query.R
\name{drop_attributes}
\alias{drop_attributes}
\title{Remove Salesforce attributes data from list}
\usage{
drop_attributes(x, object_name_append = FALSE, object_name_as_col = FALSE)
}
\arguments{
\item{x}{\code{list}; a list to be cleaned.}

\item{object_name_append}{\code{logical}; whether to include the object type
(e.g. Account or Contact) as part of the column names (e.g. Account.Name).}

\item{object_name_as_col}{\code{logical}; whether to include the object type
(e.g. Account or Contact) as a new column.}
}
\value{
\code{list} containing no 'attributes' elements.
}
\description{
This function removes elements from Salesforce data parsed to a list where
the object type and the record url persists because they were attributes on
the record and not part of the requested information.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
