% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-query.R
\name{drop_attributes_recursively}
\alias{drop_attributes_recursively}
\title{Recursively remove Salesforce attributes data from list}
\usage{
drop_attributes_recursively(x, object_name_append = FALSE)
}
\arguments{
\item{x}{\code{list}; a list to be cleaned.}
}
\value{
\code{list} containing no 'attributes' elements.
}
\description{
This function wraps the custom \code{drop_attributes} function that removes 
elements from Salesforce data parsed to a list where the object type and the 
record url persists because they were attributes on the record and not  
part of the requested information.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
