% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-query.R
\name{drop_empty_recursively}
\alias{drop_empty_recursively}
\title{Remove all NULL or zero-length elements from list}
\usage{
drop_empty_recursively(x)
}
\arguments{
\item{x}{\code{list}; a list to be cleaned.}
}
\value{
\code{list} containing no empty elements.
}
\description{
This function wraps the \code{\link[purrr:keep]{compact}} function to recursively
remove elements from lists that contain no information.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
