% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-query.R
\name{drop_nested_child_records}
\alias{drop_nested_child_records}
\title{Drop nested child records in a record}
\usage{
drop_nested_child_records(x)
}
\arguments{
\item{x}{\code{list}; a list parsed from JSON and representing one individual
parent query record.}
}
\value{
\code{list}; a list without any elements that have nested child records
assuming they have already been extracted.
}
\description{
This function accepts a single record from a nested query and removes the element
with nested "records" which represent the child records belonging to the parent.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
