% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-query.R
\name{extract_records_from_xml_node}
\alias{extract_records_from_xml_node}
\title{Pulls out a tibble of record info from an XML node}
\usage{
extract_records_from_xml_node(node, object_name_append = FALSE)
}
\arguments{
\item{node}{\code{xml_node}; the node to have records extracted into one row \code{tbl_df}.}

\item{object_name_append}{\code{logical}; an indicator for whether to append the 
name of the object onto the the front of the field names of the records.}
}
\value{
\code{tbl_df} parsed from the supplied node
}
\description{
This function accepts an \code{xml_node} and searches for all './/records' 
in the document to format into a single tidy \code{tbl_df}.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
