% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-xml.R
\name{extract_records_from_xml_nodeset_of_records}
\alias{extract_records_from_xml_nodeset_of_records}
\title{Pulls out a tibble of record info from an XML node}
\usage{
extract_records_from_xml_nodeset_of_records(x, object_name = NULL)
}
\arguments{
\item{nodeset}{\code{xml_nodeset}; nodeset to have records extracted into a \code{tbl_df}}
}
\value{
\code{tbl_df} parsed from the supplied \code{xml_nodeset}
}
\description{
This function accepts an \code{xml_nodeset} and searches for all './/records' 
in the document to format into a single tidy \code{tbl_df}.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
