% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-query.R
\name{flatten_tbl_df}
\alias{flatten_tbl_df}
\title{Flatten list and convert to tibble}
\usage{
flatten_tbl_df(x)
}
\arguments{
\item{x}{\code{list}; a list to be extracted into a \code{tbl_df}.}
}
\value{
\code{tbl_df} parsed from the flattened list.
}
\description{
This function is a convenience function to handle deeply nested records usually
returned by parsed JSON or XML that need to be converted into a data frame where
each record represents a row in the data frame.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
