% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_headers_for_verbose}
\alias{format_headers_for_verbose}
\title{Format Headers for Printing}
\usage{
format_headers_for_verbose(request_headers)
}
\description{
Format Headers for Printing
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
