% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_os}
\alias{get_os}
\title{Determine the host operating system}
\usage{
get_os()
}
\value{
\code{character}; a string indicating the current operating system.
}
\description{
This function determines whether the system running the R code
is Windows, Mac, or Linux
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\examples{
\dontrun{
get_os()
}
}
\seealso{
\url{https://conjugateprior.org/2015/06/identifying-the-os-from-r/}
}
\keyword{internal}
