% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{guess_object_name_from_soql}
\alias{guess_object_name_from_soql}
\title{Try to Guess the Object if User Does Not Specify for Bulk Queries}
\usage{
guess_object_name_from_soql(soql)
}
\arguments{
\item{soql}{\code{character}; a string defining a SOQL query
(e.g. "SELECT Id, Name FROM Account").}
}
\value{
\code{character}; a string parsed from the input that represents the
object name that the query appears to target.
}
\description{
Try to Guess the Object if User Does Not Specify for Bulk Queries
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
