% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-analytics-report-folder.R
\name{make_analytics_folder_collections_url}
\alias{make_analytics_folder_collections_url}
\title{Analytics Folder collections URL generator}
\usage{
make_analytics_folder_collections_url()
}
\description{
Analytics Folder collections URL generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
