% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-analytics-report-folder.R
\name{make_analytics_folder_operations_url}
\alias{make_analytics_folder_operations_url}
\title{Analytics Folder operations URL generator}
\usage{
make_analytics_folder_operations_url(report_folder_id)
}
\description{
Analytics Folder operations URL generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
