% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-analytics-notification.R
\name{make_analytics_notifications_limits_url}
\alias{make_analytics_notifications_limits_url}
\title{Analytics Notification limits URL generator}
\usage{
make_analytics_notifications_limits_url()
}
\description{
Analytics Notification limits URL generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
