% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-analytics-notification.R
\name{make_analytics_notifications_list_url}
\alias{make_analytics_notifications_list_url}
\title{Analytics Notification list URL generator}
\usage{
make_analytics_notifications_list_url()
}
\description{
Analytics Notification list URL generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
