% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-soap.R
\name{make_base_metadata_url}
\alias{make_base_metadata_url}
\title{Base Metadata API URL Generator}
\usage{
make_base_metadata_url()
}
\value{
\code{character}; a complete URL (as a string) that will be used to
send Metadata API calls to. This URL is specific to your instance and the API
version being used.
}
\description{
Base Metadata API URL Generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
