% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-rest.R
\name{make_base_rest_url}
\alias{make_base_rest_url}
\title{Base REST API URL Generator}
\usage{
make_base_rest_url()
}
\description{
Base REST API URL Generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
