% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{make_bulk_batches_url}
\alias{make_bulk_batches_url}
\title{Query URL Generator}
\usage{
make_bulk_batches_url(job_id, bulk_version = "2.0")
}
\description{
Query URL Generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
