% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-bulk.R
\name{make_bulk_delete_job_url}
\alias{make_bulk_delete_job_url}
\title{Bulk Delete Job Generic URL Generator}
\usage{
make_bulk_delete_job_url(job_id)
}
\description{
Bulk Delete Job Generic URL Generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
