% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-bulk.R
\name{make_bulk_end_job_generic_url}
\alias{make_bulk_end_job_generic_url}
\title{Bulk End Job Generic URL Generator}
\usage{
make_bulk_end_job_generic_url(job_id, api_type = c("Bulk 1.0", "Bulk 2.0"))
}
\description{
Bulk End Job Generic URL Generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
