% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-bulk.R
\name{make_bulk_get_all_jobs_url}
\alias{make_bulk_get_all_jobs_url}
\title{Bulk Get All Jobs Generic URL Generator}
\usage{
make_bulk_get_all_jobs_url(
  parameterized_search_list = list(isPkChunkingEnabled = NULL, jobType = NULL),
  next_records_url = NULL,
  api_type = c("Bulk 2.0")
)
}
\arguments{
\item{parameterized_search_list}{\code{list}; a list of search options to locate
Bulk API jobs.}

\item{next_records_url}{\code{character}; a string returned by a Salesforce
query from where to find subsequent records returned by a paginated query.}

\item{api_type}{\code{character}; a string indicating which Bulk API to execute
the call against.}
}
\value{
\code{character}; a complete URL (as a string) to send a request
to in order to retrieve queried jobs.
}
\description{
Bulk Get All Jobs Generic URL Generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
