% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-bulk.R
\name{make_bulk_get_job_url}
\alias{make_bulk_get_job_url}
\title{Bulk Get Job Generic URL Generator}
\usage{
make_bulk_get_job_url(
  job_id,
  api_type = c("Bulk 1.0", "Bulk 2.0"),
  query_operation = NULL
)
}
\arguments{
\item{job_id}{\code{character}; the Salesforce Id assigned to a submitted job as
returned by \link{sf_create_job_bulk}. It will start with \code{"750"}.}

\item{api_type}{\code{character}; a string indicating which Bulk API to execute
the call against.}

\item{query_operation}{\code{logical}; an indicator of whether the call is
for a query or another operation, such as, CRUD operations.}
}
\value{
\code{character}; a complete URL (as a string) that will be used to
send subsequent Bulk API calls to. This URL is specific to your instance and
the API version being used.
}
\description{
Bulk Get Job Generic URL Generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
