% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-bulk.R
\name{make_bulk_job_records_url}
\alias{make_bulk_job_records_url}
\title{Bulk Job Records URL Generator}
\usage{
make_bulk_job_records_url(job_id, record_type = c("successfulResults",
  "failedResults", "unprocessedRecords"), api_type = c("Bulk 2.0"))
}
\description{
Bulk Job Records URL Generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
