% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{make_bulk_query_url}
\alias{make_bulk_query_url}
\title{Bulk Query URL Generator}
\usage{
make_bulk_query_url(job_id)
}
\description{
Bulk Query URL Generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
