% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{make_chatter_url}
\alias{make_chatter_url}
\title{Chatter URL Generator}
\usage{
make_chatter_url()
}
\description{
Chatter URL Generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
