% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{make_create_url}
\alias{make_create_url}
\title{Create URL Generator}
\usage{
make_create_url(target_object)
}
\description{
Create URL Generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
