% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-analytics-dashboard.R
\name{make_dashboard_filter_options_analysis_url}
\alias{make_dashboard_filter_options_analysis_url}
\title{Dashboard filter options analysis URL generator}
\usage{
make_dashboard_filter_options_analysis_url(dashboard_id)
}
\description{
Dashboard filter options analysis URL generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
