% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-analytics-dashboard.R
\name{make_dashboard_status_url}
\alias{make_dashboard_status_url}
\title{Dashboard status URL generator}
\usage{
make_dashboard_status_url(dashboard_id)
}
\description{
Dashboard status URL generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
