% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-analytics-dashboard.R
\name{make_dashboards_list_url}
\alias{make_dashboards_list_url}
\title{Dashboard list URL generator}
\usage{
make_dashboards_list_url()
}
\description{
Dashboard list URL generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
