% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-rest.R
\name{make_describe_objects_url}
\alias{make_describe_objects_url}
\title{Describe Objects URL Generator}
\usage{
make_describe_objects_url(object)
}
\description{
Describe Objects URL Generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
