% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-soap.R
\name{make_login_url}
\alias{make_login_url}
\title{Login URL Generator}
\usage{
make_login_url(login_url)
}
\description{
Login URL Generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
