% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-rest.R
\name{make_parameterized_search_url}
\alias{make_parameterized_search_url}
\alias{make_parameterized_search_url}
\title{Parameterized Search URL Generator}
\usage{
make_parameterized_search_url(search_string = NULL, params = NULL)

make_parameterized_search_url(search_string = NULL, params = NULL)
}
\description{
Parameterized Search URL Generator

Parameterized Search URL Generator
}
\note{
This function is meant to be used internally. Only use when debugging.

This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
