% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-analytics-report.R
\name{make_report_copy_url}
\alias{make_report_copy_url}
\title{Report Copy URL generator}
\usage{
make_report_copy_url(report_id)
}
\description{
Report Copy URL generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
