% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-analytics-report.R
\name{make_report_create_url}
\alias{make_report_create_url}
\title{Report Create URL generator}
\usage{
make_report_create_url()
}
\value{
\code{character}; a complete URL (as a string) that will be used to
send Reports and Dashboards API calls to. This URL is specific to your instance
and the API version being used.
}
\description{
Report Create URL generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
