% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-analytics-report.R
\name{make_report_fields_url}
\alias{make_report_fields_url}
\title{Report Fields URL generator}
\usage{
make_report_fields_url(report_id)
}
\arguments{
\item{report_id}{\code{character}; the Salesforce Id assigned to a created
analytics report. It will start with \code{"00O"}.}
}
\value{
\code{character}; a complete URL (as a string) that will be used to
send Reports and Dashboards API calls to. This URL is specific to your instance
and the API version being used.
}
\description{
Report Fields URL generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
