% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-analytics-report.R
\name{make_report_filter_operators_list_url}
\alias{make_report_filter_operators_list_url}
\title{Report Filter Operator List URL generator}
\usage{
make_report_filter_operators_list_url(for_dashboards = FALSE)
}
\description{
Report Filter Operator List URL generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
