% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-analytics-report.R
\name{make_report_instance_url}
\alias{make_report_instance_url}
\title{Report Instance URL generator}
\usage{
make_report_instance_url(report_id, report_instance_id)
}
\arguments{
\item{report_id}{\code{character}; the Salesforce Id assigned to a created
analytics report. It will start with \code{"00O"}.}

\item{report_instance_id}{\code{character}; the Salesforce Id assigned to a
created analytics report instance (an asynchronous run). It will start with
\code{"0LG"}.}
}
\value{
\code{character}; a complete URL (as a string) that will be used to
send Reports and Dashboards API calls to. This URL is specific to your instance
and the API version being used.
}
\description{
Report Instance URL generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
