% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-analytics-report.R
\name{make_report_type_describe_url}
\alias{make_report_type_describe_url}
\title{Report Type Describe URL generator}
\usage{
make_report_type_describe_url(type)
}
\arguments{
\item{type}{\code{character}; The API name of a report type.}
}
\value{
\code{character}; a complete URL (as a string) that will be used to
send Reports and Dashboards API calls to. This URL is specific to your instance
and the API version being used.
}
\description{
Report Type Describe URL generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
