% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-analytics-report.R
\name{make_reports_list_url}
\alias{make_reports_list_url}
\title{Reports List URL generator}
\usage{
make_reports_list_url()
}
\description{
Reports List URL generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
