% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-rest.R
\name{make_rest_describe_url}
\alias{make_rest_describe_url}
\title{REST API Describe URL Generator}
\usage{
make_rest_describe_url(object_name)
}
\arguments{
\item{object_name}{\code{character}; the name of the Salesforce object that the
function is operating against (e.g. "Account", "Contact", "CustomObject__c").}
}
\description{
REST API Describe URL Generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
