% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-rest.R
\name{make_rest_objects_url}
\alias{make_rest_objects_url}
\title{REST Objects URL Generator}
\usage{
make_rest_objects_url(object)
}
\description{
REST Objects URL Generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
