% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest-endpoints.R
\name{make_search_url}
\alias{make_search_url}
\title{Search URL Generator}
\usage{
make_search_url(search_string)
}
\description{
Search URL Generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
