% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_soap_xml_skeleton}
\alias{make_soap_xml_skeleton}
\title{Make SOAP XML Request Skeleton}
\usage{
make_soap_xml_skeleton()
}
\arguments{
\item{input_data}{a \code{data.frame} of data to fill the XML body}

\item{externalIDFieldName}{character; a string}
}
\value{
a XML document
}
\description{
Create XML in preparate for sending to the SOAP API
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
