% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_verbose_httr_message}
\alias{make_verbose_httr_message}
\title{Format Verbose Call}
\usage{
make_verbose_httr_message(
  method,
  url,
  headers = NULL,
  body = NULL,
  auto_unbox = TRUE,
  ...
)
}
\arguments{
\item{method}{\code{character}; the type of HTTP method invoked (e.g., POST,
PUT, DELETE, etc.).}

\item{url}{\code{character}; the URL that the request was sent to}

\item{headers}{\code{character}; the set of header options set on the request}

\item{body}{\code{character}; the body of the request.}

\item{auto_unbox}{\code{logical}, an indicator of whether to parse vectors of
of length 1 into a single character string, rather than a list.}

\item{...}{additional arguments passed on to \code{\link[jsonlite]{toJSON}}.}
}
\value{
\code{NULL} invisibly, because this function is intended for the
side-effect of printing out the details of an HTTP call.
}
\description{
Format Verbose Call
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
