% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-query.R
\name{map_sf_type_to_r_type}
\alias{map_sf_type_to_r_type}
\title{Map Salesforce data types to R data types}
\usage{
map_sf_type_to_r_type(x)
}
\arguments{
\item{x}{\code{character}; the Salesforce data type.}
}
\value{
\code{character} the R data type.
}
\description{
This function is a simple one-to-many map of unique Salesforce data types to
a specific data type in R.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\seealso{
\itemize{
\item \href{https://developer.salesforce.com/docs/atlas.en-us.object_reference.meta/object_reference/primitive_data_types.htm}{Primitive Data Types}
\item \href{https://developer.salesforce.com/docs/atlas.en-us.object_reference.meta/object_reference/field_types.htm}{Other Field Types}
\item \href{https://developer.salesforce.com/docs/atlas.en-us.object_reference.meta/object_reference/compound_fields.htm}{Compound Fields}
}
}
\keyword{internal}
