% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_null_to_na}
\alias{merge_null_to_na}
\title{Return NA if NULL}
\usage{
merge_null_to_na(x)
}
\description{
Return NA if NULL
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
