% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-input-validation.R
\name{message_w_errors_listed}
\alias{message_w_errors_listed}
\title{List a vector of issues and in a message}
\usage{
message_w_errors_listed(main_text = "Consider the following:", errors = NULL)
}
\arguments{
\item{main_text}{\code{character}; The text used to introduce the list of
errors, typically ending with a colon. For example: \code{"Consider fixing the 
following issue(s):"}.}

\item{errors}{\code{errors}; a vector of issues that will be formatted into
a bulleted list for the user to review with each issue listed on a new line.}
}
\value{
\code{NULL} invisibly
}
\description{
List a vector of issues and in a message
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
