% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-httr.R
\name{parse_error_code_and_message}
\alias{parse_error_code_and_message}
\title{Function to parse out the message and status code of an HTTP error}
\usage{
parse_error_code_and_message(x)
}
\description{
Assuming the error code is less than 500, this function will return the
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
