% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_parser}
\alias{query_parser}
\title{query_parser}
\usage{
query_parser(xml)
}
\arguments{
\item{xml}{a \code{xml_document}}
}
\value{
\code{data.frame} parsed from the supplied xml
}
\description{
A function specifically for parsing SOQL query XML into data.frames
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
