% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-httr.R
\name{rDELETE}
\alias{rDELETE}
\title{DELETEs with retries and authentication}
\usage{
rDELETE(url, headers = character(0), ...)
}
\description{
DELETEs with retries and authentication
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
