% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_empty_linked_object_cols}
\alias{remove_empty_linked_object_cols}
\title{Remove NA Columns Created by Empty Related Entity Values}
\usage{
remove_empty_linked_object_cols(dat, api_type = c("SOAP", "REST"))
}
\arguments{
\item{dat}{data; a \code{tbl_df} or \code{data.frame} of a returned resultset}

\item{api_type}{character; one of "REST", "SOAP", "Bulk 1.0", "Bulk 2.0", or 
"Chatter" indicating which API to use when making the request}
}
\description{
This function will detect if there are related entity columns coming back 
in the resultset and try to exclude an additional completely blank column 
created by records that don't have a relationship at all in that related entity.
}
\keyword{internal}
