% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-control.R
\name{return_matching_controls}
\alias{return_matching_controls}
\title{Of All Args Return Ones Matching Control Arguments}
\usage{
return_matching_controls(args)
}
\description{
Of All Args Return Ones Matching Control Arguments
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
