% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{rforcecom.bulkAction}
\alias{rforcecom.bulkAction}
\title{salesforcer's backwards compatible version of rforcecom.bulkAction}
\usage{
rforcecom.bulkAction()
}
\description{
salesforcer's backwards compatible version of rforcecom.bulkAction
}
