% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{rforcecom.bulkQuery}
\alias{rforcecom.bulkQuery}
\title{salesforcer's backwards compatible version of rforcecom.bulkQuery}
\usage{
rforcecom.bulkQuery()
}
\description{
salesforcer's backwards compatible version of rforcecom.bulkQuery
}
