% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{rforcecom.create}
\alias{rforcecom.create}
\title{salesforcer's backwards compatible version of rforcecom.create}
\usage{
rforcecom.create()
}
\description{
salesforcer's backwards compatible version of rforcecom.create
}
