% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{rforcecom.delete}
\alias{rforcecom.delete}
\title{salesforcer's backwards compatible version of rforcecom.delete}
\usage{
rforcecom.delete()
}
\description{
salesforcer's backwards compatible version of rforcecom.delete
}
