% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{rforcecom.getObjectDescription}
\alias{rforcecom.getObjectDescription}
\title{salesforcer's backwards compatible version of rforcecom.getObjectDescription}
\usage{
rforcecom.getObjectDescription(session, objectName)
}
\arguments{
\item{session}{\code{list}; a list containing "sessionID", "instanceURL", and "
apiVersion" as returned by \code{RForcecom::rforcecom.login()}. This argument is
ignored in all backward compatible calls because the authorization credentials 
are stored in an environment internal to the salesforcer package, so it is no longer 
necessary to pass the session in each function call.}

\item{objectName}{character; the name of the Salesforce object that the 
function is operating against (e.g. "Account", "Contact", "CustomObject__c")}
}
\value{
Object descriptions
}
\description{
salesforcer's backwards compatible version of rforcecom.getObjectDescription
}
\note{
This function returns a data.frame with one row per field for an object.
}
