% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{rforcecom.getServerTimestamp}
\alias{rforcecom.getServerTimestamp}
\title{salesforcer's backwards compatible version of rforcecom.getServerTimestamp}
\usage{
rforcecom.getServerTimestamp()
}
\description{
salesforcer's backwards compatible version of rforcecom.getServerTimestamp
}
