% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{rforcecom.query}
\alias{rforcecom.query}
\title{salesforcer's backwards compatible version of rforcecom.query}
\usage{
rforcecom.query(session, soqlQuery, queryAll = FALSE)
}
\arguments{
\item{session}{Session data}

\item{soqlQuery}{character; a valid SOQL string}

\item{queryAll}{logical; indicating if the query recordset should include 
deleted and archived records (available only when querying Task and Event records)}
}
\description{
salesforcer's backwards compatible version of rforcecom.query
}
