% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{rforcecom.retrieve}
\alias{rforcecom.retrieve}
\title{salesforcer's backwards compatible version of rforcecom.retrieve}
\usage{
rforcecom.retrieve()
}
\description{
salesforcer's backwards compatible version of rforcecom.retrieve
}
