% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{rforcecom.search}
\alias{rforcecom.search}
\title{salesforcer's backwards compatible version of rforcecom.search}
\usage{
rforcecom.search()
}
\description{
salesforcer's backwards compatible version of rforcecom.search
}
