% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{rforcecom.update}
\alias{rforcecom.update}
\title{salesforcer's backwards compatible version of rforcecom.update}
\usage{
rforcecom.update()
}
\description{
salesforcer's backwards compatible version of rforcecom.update
}
