% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{rforcecom.upsert}
\alias{rforcecom.upsert}
\title{salesforcer's backwards compatible version of rforcecom.upsert}
\usage{
rforcecom.upsert()
}
\description{
salesforcer's backwards compatible version of rforcecom.upsert
}
