% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-httr.R
\name{rGET}
\alias{rGET}
\title{GETs with retries and authentication}
\usage{
rGET(url, headers = character(0), ...)
}
\description{
GETs with retries and authentication
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
