% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-httr.R
\name{rPATCH}
\alias{rPATCH}
\title{PATCHs with retries and authentication}
\usage{
rPATCH(url, headers = character(0), ...)
}
\description{
PATCHs with retries and authentication
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
