% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-httr.R
\name{rPOST}
\alias{rPOST}
\title{POSTs with retries and authentication}
\usage{
rPOST(url, headers = character(0), ...)
}
\description{
POSTs with retries and authentication
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
