% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-httr.R
\name{rPUT}
\alias{rPUT}
\title{PUTs with retries and authentication}
\usage{
rPUT(url, headers = character(0), ...)
}
\description{
PUTs with retries and authentication
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
