% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salesforcer.R
\docType{package}
\name{salesforcer}
\alias{salesforcer}
\alias{salesforcer-package}
\title{\code{salesforcer} package}
\description{
An R package connecting to Salesforce APIs using tidy principles
}
\details{
A package that connects to Salesforce via REST, SOAP, Bulk, and Metadata APIs 
and emphasizes the use of tidy data principles and the tidyverse.

Additional material can be found in the 
\href{https://github.com/reportmort/salesforcer}{README} on GitHub
}
