% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{salesforcer_state}
\alias{salesforcer_state}
\title{Return the package's .state environment variable}
\usage{
salesforcer_state()
}
\value{
\code{list}; a list of values stored in the package's .state environment variable
}
\description{
Return the package's .state environment variable
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
