% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{session_id_available}
\alias{session_id_available}
\title{Check session_id availability}
\usage{
session_id_available(verbose = TRUE)
}
\value{
\code{logical}
}
\description{
Check if a session_id is available in \code{\link{salesforcer}}'s internal
\code{.state} environment.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
