% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-query.R
\name{set_null_elements_to_na_recursively}
\alias{set_null_elements_to_na_recursively}
\title{Recursively set all NULL or zero-length elements from list to NA}
\usage{
set_null_elements_to_na_recursively(x)
}
\arguments{
\item{x}{\code{list}; a list to be cleaned.}
}
\value{
\code{list} containing \code{NA} in place of \code{NULL} element values.
}
\description{
This function wraps a simple \code{\link[purrr:modify]{modify_if}} function
to recursively set NULL elements in a list to NA.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
