% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{sf_access_token}
\alias{sf_access_token}
\title{Return access_token attribute of OAuth 2.0 Token}
\usage{
sf_access_token(verbose = FALSE)
}
\arguments{
\item{verbose}{\code{logical}; an indicator of whether to print additional
detail for each API call, which is useful for debugging. More specifically, when
set to \code{TRUE} the URL, header, and body will be printed for each request,
along with additional diagnostic information where available.}
}
\value{
\code{character}; a string of the access_token element of the current token in
force; otherwise NULL
}
\description{
Return access_token attribute of OAuth 2.0 Token
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
