% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-notification.R
\name{sf_analytics_notification_create}
\alias{sf_analytics_notification_create}
\title{Create an analytics notification}
\usage{
sf_analytics_notification_create(body)
}
\description{
Create an analytics notification
}
