% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-notification.R
\name{sf_analytics_notification_delete}
\alias{sf_analytics_notification_delete}
\title{Delete an analytics notification}
\usage{
sf_analytics_notification_delete(notification_id)
}
\description{
Delete an analytics notification
}
