% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-notification.R
\name{sf_analytics_notification_describe}
\alias{sf_analytics_notification_describe}
\title{Describe an analytics notification}
\usage{
sf_analytics_notification_describe(notification_id)
}
\description{
Describe an analytics notification
}
