% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-notification.R
\name{sf_analytics_notification_update}
\alias{sf_analytics_notification_update}
\title{Update an analytics notification}
\usage{
sf_analytics_notification_update(notification_id, body)
}
\description{
Update an analytics notification
}
