% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-notification.R
\name{sf_analytics_notifications_limits}
\alias{sf_analytics_notifications_limits}
\title{Return limits of analytics notifications}
\usage{
sf_analytics_notifications_limits(
  source = c("lightningDashboardSubscribe", "lightningReportSubscribe",
    "waveNotification"),
  record_id = NULL
)
}
\description{
Return limits of analytics notifications
}
