% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-notification.R
\name{sf_analytics_notifications_list}
\alias{sf_analytics_notifications_list}
\title{List analytics notifications}
\usage{
sf_analytics_notifications_list(
  source = c("lightningDashboardSubscribe", "lightningReportSubscribe",
    "waveNotification"),
  owner_id = NULL,
  record_id = NULL
)
}
\description{
List analytics notifications
}
