% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{sf_auth_refresh}
\alias{sf_auth_refresh}
\title{Refresh an existing Authorized Salesforce Token}
\usage{
sf_auth_refresh(verbose = FALSE)
}
\arguments{
\item{verbose}{\code{logical}; an indicator of whether to print additional
detail for each API call, which is useful for debugging. More specifically, when
set to \code{TRUE} the URL, header, and body will be printed for each request,
along with additional diagnostic information where available.}
}
\value{
a \code{Token2.0} object (an S3 class provided by \code{httr}) or a
a character string of the sessionId element of the current authorized
API session
}
\description{
Force the current OAuth to refresh. This is only needed for times when you
load the token from outside the current working directory, it is expired, and
you're running in non-interactive mode.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
