% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-query.R
\name{sf_build_cols_spec}
\alias{sf_build_cols_spec}
\title{Produce spec to convert Salesforce data types to R data types}
\usage{
sf_build_cols_spec(x)
}
\arguments{
\item{x}{\code{character}; the Salesforce data types to map}
}
\value{
\code{character} the analogous R data types.
}
\description{
This function accepts a vector of Salesforce data types and maps them into
a single string that can be passed to the \code{col_types} argument.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
