% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk-operation.R
\name{sf_bulk_batch_details}
\alias{sf_bulk_batch_details}
\title{Returning the Details of a Batch in a Bulk API Job}
\usage{
sf_bulk_batch_details(job_id, batch_id)
}
\arguments{
\item{job_id}{character; the Salesforce Id assigned to a submitted job as 
returned by \link{sf_bulk_create_job}}

\item{batch_id}{character; the Salesforce Id assigned assigned to a submitted 
batch as returned by \link{sf_bulk_create_batches}}
}
\value{
A \code{tbl_df}, formatted by salesforce, with information containing the success or failure or certain rows in a submitted batch, 
unless the operation was query, then it is a data.frame containing the result_id for retrieving the recordset.
}
\description{
This function returns detailed (row-by-row) information on an existing batch
which has already been submitted to Bulk API Job
}
\note{
This is typically a legacy function used only with Bulk 1.0, but many 
operations such as insert, update, upsert, delete have been moved to using Bulk 2.0.
}
\examples{
\dontrun{
job_info <- sf_bulk_create_job(operation = "query", object = "Account")
soql <- "SELECT Id, Name FROM Account LIMIT 10"
batch_query_info <- sf_bulk_submit_query(job_id = job_info$id, soql = soql)
batch_details <- sf_bulk_batch_details(job_id=batch_query_info$jobId,
                                       batch_id=batch_query_info$id)
sf_bulk_close_job(job_info$id)
}
}
\references{
\url{https://developer.salesforce.com/docs/atlas.en-us.api_asynch.meta/api_asynch/}
}
