% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk-operation.R
\name{sf_bulk_close_job}
\alias{sf_bulk_close_job}
\title{Close Bulk API Job}
\usage{
sf_bulk_close_job(job_id)
}
\arguments{
\item{job_id}{character; the Salesforce Id assigned to a submitted job as 
returned by \link{sf_bulk_create_job}}
}
\value{
A \code{list} of parameters defining the now closed job
}
\description{
This function closes a Job in the Salesforce Bulk API
}
\note{
This is typically a legacy function used only with Bulk 1.0. It is used less 
frequently (only with bulk queries) since the operations insert, update, upsert, delete 
have all been moved to using Bulk 2.0.
}
\examples{
\dontrun{
my_query <- "SELECT Id, Name FROM Account LIMIT 10"
job_info <- sf_bulk_create_job(operation='query', object='Account')
query_info <- sf_bulk_submit_query(job_id=job_info$id, soql=my_query)
recordset <- sf_bulk_query_result(job_id = query_info$jobId,
                                  batch_id = query_info$id,
                                  result_id = result$result)
sf_bulk_close_job(job_info$id)
}
}
\references{
\url{https://developer.salesforce.com/docs/atlas.en-us.api_asynch.meta/api_asynch/}
}
