% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk-operation.R
\name{sf_bulk_create_batches}
\alias{sf_bulk_create_batches}
\title{Add Batches to a Bulk API Job}
\usage{
sf_bulk_create_batches(job_id, input_data)
}
\arguments{
\item{job_id}{character; the Salesforce Id assigned to a submitted job as 
returned by \link{sf_bulk_create_job}}

\item{input_data}{\code{named vector}, \code{matrix}, \code{data.frame}, or 
\code{tbl_df}; data can be coerced into .csv file for submitting as batch request}
}
\value{
A \code{list} of \code{list}s, one for each batch submitted, containing 10 parameters of the batch
}
\description{
This function takes a data frame and submits it in batches to a 
an already existing Bulk API Job by chunking into temp files
}
\examples{
\dontrun{
# NOTE THAT YOU MUST FIRST CREATE AN EXTERNAL ID FIELD CALLED My_External_Id 
# BEFORE RUNNING THIS EXAMPLE
# inserting 2 records
my_data <- tibble(Name=c('New Record 1', 'New Record 2'),
                  My_External_Id__c=c('11111','22222'))
job_info <- sf_bulk_create_job(operation='insert',
                               object='Account')
batches_ind <- sf_bulk_create_batches(job_id = job_info$id,
                                      input_data = my_data)
# upserting 3 records
my_data2 <- tibble(My_External_Id__c=c('11111','22222', '99999'), 
                  Name=c('Updated_Name1', 'Updated_Name2', 'Upserted_Record')) 
job_info <- sf_bulk_create_job(operation='upsert',
                               externalIdFieldName='My_External_Id__c',
                               object='Account')
batches_ind <- sf_bulk_create_batches(job_id = job_info$id,
                                      input_data = my_data2)
sf_bulk_get_job(job_info$id)                                     
}
}
\references{
\url{https://developer.salesforce.com/docs/atlas.en-us.api_asynch.meta/api_asynch/}
}
