% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk-operation.R
\name{sf_bulk_delete_job}
\alias{sf_bulk_delete_job}
\title{Delete Bulk API Job}
\usage{
sf_bulk_delete_job(job_id)
}
\arguments{
\item{job_id}{character; the Salesforce Id assigned to a submitted job as 
returned by \link{sf_bulk_create_job}}
}
\description{
Delete Bulk API Job
}
\examples{
\dontrun{
job_info <- sf_bulk_create_job('insert', 'Account')
sf_bulk_abort_job(job_info$id)
sf_bulk_delete_job(job_info$id)
}
}
