% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk-operation.R
\name{sf_bulk_end_job_generic}
\alias{sf_bulk_end_job_generic}
\title{End Bulk API Job}
\usage{
sf_bulk_end_job_generic(job_id, end_type = c("Closed", "UploadComplete",
  "Aborted"), bulk_version = c("2.0", "1.0"))
}
\arguments{
\item{job_id}{character; the Salesforce Id assigned to a submitted job as 
returned by \link{sf_bulk_create_job}}

\item{end_type}{character; taking a value of "Closed" or "Aborted" indicating 
how the bulk job should be ended}

\item{bulk_version}{character; taking a value of "2.0" or "1.0" indicating which 
version of Bulk API is being used}
}
\description{
End Bulk API Job
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
