% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk-operation.R
\name{sf_bulk_get_job}
\alias{sf_bulk_get_job}
\title{Get Bulk API Job}
\usage{
sf_bulk_get_job(job_id)
}
\arguments{
\item{job_id}{character; the Salesforce Id assigned to a submitted job as 
returned by \link{sf_bulk_create_job}}
}
\value{
A \code{list} of parameters defining the details of the specified job id
}
\description{
This function retrieves details about a Job in the Salesforce Bulk API
}
\examples{
\dontrun{
job_info <- sf_bulk_create_job('insert', 'Account')
refreshed_job_info <- sf_bulk_get_job(job_info$id)
sf_bulk_abort_job(refreshed_job_info$id)
}
}
\references{
\url{https://developer.salesforce.com/docs/atlas.en-us.api_asynch.meta/api_asynch/}
}
