% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk-operation.R
\name{sf_bulk_get_job_records}
\alias{sf_bulk_get_job_records}
\title{Returning the Details of a Bulk API Job}
\usage{
sf_bulk_get_job_records(job_id, record_types = c("successfulResults",
  "failedResults", "unprocessedRecords"))
}
\arguments{
\item{job_id}{character; the Salesforce Id assigned to a submitted job as 
returned by \link{sf_bulk_create_job}}

\item{record_types}{character; one or more types of records to retrieve from 
the results of running the specified job}
}
\value{
A \code{list} of \code{tbl_df}, formatted by Salesforce, with information 
containing the success or failure or certain rows in a submitted job
}
\description{
This function returns detailed (row-by-row) information on a job
which has already been submitted completed (successfully or not).
}
\examples{
\dontrun{
job_info <- sf_bulk_create_job('insert', 'Account')
input_data <- tibble(Name=c("Test Account 1", "Test Account 2"))
batches_result <- sf_bulk_create_batches(job_info$id, input_data)
# pause 1 second for operation to finish, wait longer if job is not complete
Sys.sleep(1)
job_status <- sf_bulk_get_job(job_info$id)
if(job_status$state == "JobComplete"){
  job_record_details <- sf_bulk_get_job_records(job_id=job_info$id)
}
}
}
\references{
\url{https://developer.salesforce.com/docs/atlas.en-us.api_asynch.meta/api_asynch/}
}
