% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk-query.R
\name{sf_bulk_query}
\alias{sf_bulk_query}
\title{Run Bulk Query}
\usage{
sf_bulk_query(soql, object, interval_seconds = 5, max_attempts = 100,
  verbose = FALSE)
}
\arguments{
\item{soql}{character; a string defining a SOQL query. (ex: "SELECT Id, Name FROM Account")}

\item{object}{a character string defining the target salesforce object that the operation will be performed on. 
This must match the target object in the query}

\item{interval_seconds}{an integer defining the seconds between attempts to check for job completion}

\item{max_attempts}{an integer defining then max number attempts to check for job completion before stopping}

\item{verbose}{logical; do you want informative messages?}
}
\value{
A \code{data.frame} of the recordset returned by query
}
\description{
This function is a convenience wrapper for submitting and retrieving 
bulk query API jobs
}
\examples{
\dontrun{
# select all Ids from Account object
ids <- sf_bulk_query(soql='Select Id from Account', object='Account')
}
}
\references{
\url{https://developer.salesforce.com/docs/atlas.en-us.api_asynch.meta/api_asynch/}
}
