% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk-query.R
\name{sf_bulk_query_result}
\alias{sf_bulk_query_result}
\title{Retrieving the Results of a Bulk Query Batch in a Bulk API Job}
\usage{
sf_bulk_query_result(job_id, batch_id, result_id)
}
\arguments{
\item{job_id}{character; the Salesforce Id assigned to a submitted job as 
returned by \link{sf_bulk_create_job}}

\item{batch_id}{character; the Salesforce Id assigned assigned to a submitted 
batch as returned by \link{sf_bulk_create_batches}}

\item{result_id}{a character string returned from \link{sf_bulk_batch_details} when a query has completed and specifies how to get the recordset}
}
\value{
A \code{tbl_df}, formatted by salesforce, containing query results
}
\description{
This function returns the resultset of a Bulk Query batch
which has already been submitted to Bulk API Job and has Completed state
}
\examples{
\dontrun{
my_query <- "SELECT Id, Name FROM Account LIMIT 10"
job_info <- sf_bulk_create_job(operation='query', object='Account')
query_info <- sf_bulk_submit_query(job_id=job_info$id, soql=my_query)
result <- sf_bulk_batch_details(job_id = query_info$jobId,
                                batch_id = query_info$id)
recordset <- sf_bulk_query_result(job_id = query_info$jobId,
                                  batch_id = query_info$id,
                                  result_id = result$result)
sf_bulk_close_job(job_info$id)
}
}
\references{
\url{https://developer.salesforce.com/docs/atlas.en-us.api_asynch.meta/api_asynch/}
}
