% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk-query.R
\name{sf_bulk_submit_query}
\alias{sf_bulk_submit_query}
\title{Submit Bulk Query Batch to a Bulk API Job}
\usage{
sf_bulk_submit_query(job_id, soql)
}
\arguments{
\item{job_id}{character; the Salesforce Id assigned to a submitted job as 
returned by \link{sf_bulk_create_job}}

\item{soql}{a character string defining a valid SOQL query on the Salesforce object associated with the job}
}
\value{
A \code{list} parameters of the batch
}
\description{
This function takes a SOQL text string and submits the query to 
an already existing Bulk API Job of operation "query"
}
\note{
Bulk API query doesn't support the following SOQL:
\itemize{
   \item COUNT
   \item ROLLUP
   \item SUM
   \item GROUP BY CUBE
   \item OFFSET
   \item Nested SOQL queries
   \item Relationship fields
   }
Additionally, Bulk API can't access or query compound address or compound geolocation fields.
}
\examples{
\dontrun{
my_query <- "SELECT Id, Name FROM Account LIMIT 10"
job_info <- sf_bulk_create_job(operation='query', object='Account')
query_info <- sf_bulk_submit_query(job_id=job_info$id, soql=my_query)
}
}
\references{
\url{https://developer.salesforce.com/docs/atlas.en-us.api_asynch.meta/api_asynch/}
}
