% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk-operation.R
\name{sf_bulk_upload_complete}
\alias{sf_bulk_upload_complete}
\title{Signal Upload Complete to Bulk API Job}
\usage{
sf_bulk_upload_complete(job_id)
}
\arguments{
\item{job_id}{character; the Salesforce Id assigned to a submitted job as 
returned by \link{sf_bulk_create_job}}
}
\value{
A \code{list} of parameters defining the job after signaling a completed upload
}
\description{
This function signals that uploads are complete to a Job in the Salesforce Bulk API
}
\note{
This function is typically not used directly. It is used in \code{sf_bulk_create_batches()} 
right after submitting the batches to signal to Salesforce that the batches should 
no longer be queued.
}
\examples{
\dontrun{
upload_info <- sf_bulk_upload_complete(job_id=job_info$id)
}
}
\references{
\url{https://developer.salesforce.com/docs/atlas.en-us.api_asynch.meta/api_asynch/}
}
