% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachments.R
\name{sf_create_attachment_bulk_v1}
\alias{sf_create_attachment_bulk_v1}
\title{Create Attachments using Bulk 1.0 API}
\usage{
sf_create_attachment_bulk_v1(
  input_data,
  object_name = "Attachment",
  content_type = "ZIP_CSV",
  control,
  ...,
  verbose = FALSE
)
}
\description{
Create Attachments using Bulk 1.0 API
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
