% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachments.R
\name{sf_create_attachment_soap}
\alias{sf_create_attachment_soap}
\title{Create Attachment using SOAP API}
\usage{
sf_create_attachment_soap(
  input_data,
  object_name = "Attachment",
  control,
  ...,
  verbose = FALSE
)
}
\description{
Create Attachment using SOAP API
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
