% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{sf_create_bulk_v1}
\alias{sf_create_bulk_v1}
\title{Create Records using Bulk 1.0 API}
\usage{
sf_create_bulk_v1(
  input_data,
  object_name,
  guess_types = TRUE,
  control,
  ...,
  verbose = FALSE
)
}
\description{
Create Records using Bulk 1.0 API
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
