% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk-operation.R
\name{sf_create_job_bulk_v1}
\alias{sf_create_job_bulk_v1}
\title{Create Job using Bulk 1.0 API}
\usage{
sf_create_job_bulk_v1(
  operation = c("insert", "delete", "upsert", "update", "hardDelete", "query",
    "queryall"),
  object_name,
  external_id_fieldname = NULL,
  content_type = c("CSV", "ZIP_CSV", "ZIP_XML", "ZIP_JSON"),
  concurrency_mode = c("Parallel", "Serial"),
  control,
  ...,
  verbose = FALSE
)
}
\description{
Create Job using Bulk 1.0 API
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
