% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk-operation.R
\name{sf_create_job_bulk_v2}
\alias{sf_create_job_bulk_v2}
\title{Create Job using Bulk 2.0 API}
\usage{
sf_create_job_bulk_v2(
  operation = c("insert", "delete", "upsert", "update"),
  object_name,
  external_id_fieldname = NULL,
  content_type = "CSV",
  column_delimiter = c("COMMA", "TAB", "PIPE", "SEMICOLON", "CARET", "BACKQUOTE"),
  control,
  ...,
  verbose = FALSE
)
}
\description{
Create Job using Bulk 2.0 API
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
