% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{sf_create_rest}
\alias{sf_create_rest}
\title{Create Records using REST API}
\usage{
sf_create_rest(input_data, object_name, control, ..., verbose = FALSE)
}
\description{
Create Records using REST API
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
