% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{sf_create_soap}
\alias{sf_create_soap}
\title{Create Records using SOAP API}
\usage{
sf_create_soap(input_data, object_name, control, ..., verbose = FALSE)
}
\description{
Create Records using SOAP API
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
