% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-dashboard.R
\name{sf_dashboard_components_describe}
\alias{sf_dashboard_components_describe}
\title{Describe dashboard components}
\usage{
sf_dashboard_components_describe(dashboard_id, component_ids = c(character(0)))
}
\description{
Describe dashboard components
}
