% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-dashboard.R
\name{sf_dashboard_delete}
\alias{sf_dashboard_delete}
\title{Delete a dashboard}
\usage{
sf_dashboard_delete(dashboard_id)
}
\arguments{
\item{dashboard_id}{\code{character}; the Salesforce Id assigned to a created
dashboard. It will start with \code{"01Z"}.}
}
\value{
\code{logical}
}
\description{
\lifecycle{experimental}
}
