% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-dashboard.R
\name{sf_dashboard_describe}
\alias{sf_dashboard_describe}
\title{Describe a dashboard}
\usage{
sf_dashboard_describe(dashboard_id)
}
\arguments{
\item{dashboard_id}{\code{character}; the Salesforce Id assigned to a created 
dashboard. It will start with \code{"01Z"}.}
}
\value{
\code{list}
}
\description{
Describe a dashboard
}
