% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-dashboard.R
\name{sf_dashboard_filter_operators_list}
\alias{sf_dashboard_filter_operators_list}
\title{List dashboard filter operators}
\usage{
sf_dashboard_filter_operators_list()
}
\value{
\code{list}
}
\description{
\lifecycle{experimental}
}
