% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-dashboard.R
\name{sf_dashboard_filter_options_analysis}
\alias{sf_dashboard_filter_options_analysis}
\title{Get an analysis of the filter options for a dashboard}
\usage{
sf_dashboard_filter_options_analysis(
  dashboard_id,
  filter_columns = list(),
  options = list()
)
}
\description{
Get an analysis of the filter options for a dashboard
}
