% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-dashboard.R
\name{sf_dashboard_refresh}
\alias{sf_dashboard_refresh}
\title{Refresh an existing dashboard}
\usage{
sf_dashboard_refresh(dashboard_id, filters = list())
}
\description{
Refresh an existing dashboard
}
