% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-dashboard.R
\name{sf_dashboard_results}
\alias{sf_dashboard_results}
\title{Get the results of an existing dashboard}
\usage{
sf_dashboard_results(dashboard_id, running_user = NULL, filters = list())
}
\description{
This function allows for pulling specific data from a dashboard. There is a
convenience function (\link{sf_get_dashboard_data}) to get the dashboard data
in a tabular format returned as a \code{tbl_df}.
}
