% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-dashboard.R
\name{sf_dashboard_set_sticky_filter}
\alias{sf_dashboard_set_sticky_filter}
\title{Set a sticky dashboard filter}
\usage{
sf_dashboard_set_sticky_filter(dashboard_id, filters = list())
}
\description{
Set a default filter value which gets applied to a dashboard when you open
it. The default filter value you specify only applies to you (other people
won’t see it when they open the dashboard). If you change the filter value
while viewing the dashboard, then the filter value you set in the user
interface overwrites the value you set via the API. To set sticky filters for
a dashboard, \code{canUseStickyFilter} must equal true.
}
