% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-dashboard.R
\name{sf_dashboard_status}
\alias{sf_dashboard_status}
\title{Get the status of a dashboard}
\usage{
sf_dashboard_status(dashboard_id, running_user = NULL, filters = list())
}
\description{
Get the status of a dashboard
}
