% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-dashboard.R
\name{sf_dashboard_update}
\alias{sf_dashboard_update}
\title{Update a dashboard}
\usage{
sf_dashboard_update(dashboard_id, body)
}
\arguments{
\item{dashboard_id}{\code{character}; the Salesforce Id assigned to a created
dashboard. It will start with \code{"01Z"}.}

\item{body}{\code{list}; a list of data to be formatted as JSON before being
sent to an API endpoint. This is NOT validated and is only experimental in
allowing users to utilize the API functionality before wrapper functions
have been written in R to format and validate the request in a more
user-friendly format.}
}
\value{
\code{list}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
