% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{sf_delete}
\alias{sf_delete}
\title{Delete Records}
\usage{
sf_delete(ids, object, all_or_none = FALSE, api_type = c("REST", "SOAP",
  "Bulk", "Async"), verbose = FALSE)
}
\arguments{
\item{ids}{\code{vector}, \code{matrix}, \code{data.frame}, or 
\code{tbl_df}; if not a vector, there must be a column called Id (case-insensitive) 
that can be passed in the request}

\item{object}{character; the name of the Salesforcer object that the 
function is operating against}

\item{all_or_none}{logical; allows a call to roll back all changes unless all 
records are processed successfully.}

\item{api_type}{character; one of "REST", "SOAP", "Bulk", or "Async" indicating 
which API to use when making the request}

\item{verbose}{logical; do you want informative messages?}
}
\value{
\code{tibble}
}
\description{
Deletes one or more new records to your organization’s data.
}
\examples{
\dontrun{
contacts_to_delete <- tibble(Id = c("003x000000xxxxxZZZ", "003x000000xxxxxZZZ"))
sf_delete(contacts_to_delete, object="Contact") 

contacts_to_delete <- c("003x000000xxxxxZZZ", "003x000000xxxxxZZZ")
sf_delete(contacts_to_delete, object="Contact")   

contacts_to_delete <- c("003x000000xxxxxZZZ", "003x000000xxxxxZZZ")
sf_delete(contacts_to_delete, object="Contact", api_type="Bulk")                             
}
}
