% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{sf_delete_bulk_v2}
\alias{sf_delete_bulk_v2}
\title{Delete records using Bulk 2.0 API}
\usage{
sf_delete_bulk_v2(
  ids,
  object_name,
  guess_types = TRUE,
  control,
  ...,
  verbose = FALSE
)
}
\description{
Delete records using Bulk 2.0 API
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
