% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk-operation.R
\name{sf_delete_job_bulk}
\alias{sf_delete_job_bulk}
\title{Delete Bulk API Job}
\usage{
sf_delete_job_bulk(job_id, api_type = c("Bulk 2.0"), verbose = FALSE)
}
\arguments{
\item{job_id}{character; the Salesforce Id assigned to a submitted job as 
returned by \link{sf_create_job_bulk}}

\item{api_type}{character; one of "REST", "SOAP", or "Bulk" indicating 
which API to use when making the request}

\item{verbose}{logical; do you want informative messages?}
}
\description{
Delete Bulk API Job
}
\examples{
\dontrun{
job_info <- sf_create_job_bulk('insert', 'Account')
sf_abort_job_bulk(job_info$id)
sf_delete_job_bulk(job_info$id)
}
}
