% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{sf_delete_soap}
\alias{sf_delete_soap}
\title{Delete records using SOAP API}
\usage{
sf_delete_soap(
  ids,
  object_name = NULL,
  guess_types = TRUE,
  control,
  ...,
  verbose = FALSE
)
}
\description{
Delete records using SOAP API
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
