% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-dashboard.R
\name{sf_describe_dashboard_components}
\alias{sf_describe_dashboard_components}
\title{Describe dashboard components}
\usage{
sf_describe_dashboard_components(dashboard_id, component_ids = c(character(0)))
}
\arguments{
\item{dashboard_id}{\code{character}; the Salesforce Id assigned to a created
dashboard. It will start with \code{"01Z"}.}

\item{component_ids}{\code{character}; a vector of Unique Salesforce Ids of a
dashboard component.}
}
\value{
\code{list}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
