% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe-metadata.R
\name{sf_describe_metadata}
\alias{sf_describe_metadata}
\title{Describe the Metadata in an Organization}
\usage{
sf_describe_metadata(verbose = FALSE)
}
\arguments{
\item{verbose}{logical; do you want informative messages?}
}
\value{
A \code{tbl_df}
}
\description{
This function returns details about the organization metadata
}
\examples{
\dontrun{
# describe metadata for the organization associated with the session
metadata_info <- sf_describe_metadata()
}
}
\references{
\url{https://developer.salesforce.com/docs/atlas.en-us.api_meta.meta/api_meta/}
}
