% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{sf_describe_objects}
\alias{sf_describe_objects}
\title{SObject Basic Information}
\usage{
sf_describe_objects(objects, api_type = c("SOAP", "REST", "Bulk"),
  verbose = FALSE)
}
\arguments{
\item{objects}{character; the name of one or more Salesforcer objects that the 
function is operating against (e.g. "Account", "Contact", "CustomObject__c")}

\item{api_type}{character; one of "REST", "SOAP", "Bulk", or "Async" indicating 
which API to use when making the request}

\item{verbose}{logical; do you want informative messages?}
}
\value{
\code{tibble}
}
\description{
Describes the individual metadata for the specified object.
}
\examples{
\dontrun{
account_metadata <- sf_describe_objects("Account")
account_metadata_SOAP <- sf_describe_objects("Account", api_type="SOAP")
multiple_objs_metadata <- sf_describe_objects(c("Contact", "Lead"))
}
}
