% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-dashboard.R
\name{sf_filter_dashboard_options_analysis}
\alias{sf_filter_dashboard_options_analysis}
\title{Get an analysis of the filter options for a dashboard}
\usage{
sf_filter_dashboard_options_analysis(
  dashboard_id,
  filter_columns = list(),
  options = list()
)
}
\arguments{
\item{dashboard_id}{\code{character}; the Salesforce Id assigned to a created
dashboard. It will start with \code{"01Z"}.}

\item{filter_columns}{\code{list}; a list of fields from the source report
which you check filter values against. Each object in the array has these
properties:...}

\item{options}{\code{list}; a list of objects describing a dashboard filter.
Each object has these properties:...}
}
\value{
\code{list}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
