% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datatypes.R
\name{sf_format_time.list}
\alias{sf_format_time.list}
\title{Format all Date and Datetime columns in a list}
\usage{
\method{sf_format_time}{list}(x)
}
\description{
Format all Date and Datetime columns in a list
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\seealso{
\url{https://developer.salesforce.com/docs/atlas.en-us.api_asynch.meta/api_asynch/datafiles_date_format.htm}
}
\keyword{internal}
