% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datatypes.R
\name{sf_format_time}
\alias{sf_format_time}
\title{Format all Date and Datetime columns in a list}
\usage{
sf_format_time(x, ...)
}
\arguments{
\item{x}{data which may or may not have values, elements, columns that
represent a datetime. If so, each of those are cast to the ISO8601 standard
per the requirements of Salesforce APIs.}
}
\value{
the same data object with datetime values formatted.
}
\description{
Format all Date and Datetime columns in a list
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\seealso{
\url{https://developer.salesforce.com/docs/atlas.en-us.api_asynch.meta/api_asynch/datafiles_date_format.htm}
}
\keyword{internal}
