% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-dashboard.R
\name{sf_get_dashboard_data}
\alias{sf_get_dashboard_data}
\title{Get dashboard data in a tabular format}
\usage{
sf_get_dashboard_data(dashboard_id, running_user = NULL, filters = list())
}
\description{
Get dashboard data in a tabular format
}
