% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-report.R
\name{sf_get_report_data}
\alias{sf_get_report_data}
\title{Get a report's data in tabular format}
\usage{
sf_get_report_data(report_id)
}
\arguments{
\item{report_id}{\code{character}; the Salesforce Id assigned to a created 
analytics report. It will start with \code{"00O"}.}
}
\value{
\code{tbl_df}
}
\description{
Get a report's data in tabular format
}
