% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-query.R
\name{sf_guess_cols}
\alias{sf_guess_cols}
\title{Parse resultset columns to a known datatype in R}
\usage{
sf_guess_cols(df, guess_types = TRUE, dataType = NULL)
}
\arguments{
\item{df}{\code{tbl_df}; the data frame to rearrange columns in}
}
\value{
\code{tbl_df} the formatted data frame
}
\description{
This function accepts a \code{tbl_df} with columns rearranged.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
