% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-org.R
\name{sf_list_api_limits}
\alias{sf_list_api_limits}
\title{List the Limits for an API}
\usage{
sf_list_api_limits()
}
\value{
\code{list}
}
\description{
\lifecycle{stable}

Lists information about limits in your org.
}
\note{
This resource is available in REST API version 29.0 and later for API
users with the View Setup and Configuration permission. The resource returns
these limits:
\itemize{
\item Daily API calls
\item Daily asynchronous Apex method executions (batch Apex, future methods, queueable Apex, and scheduled Apex)
\item Daily Bulk API calls
\item Daily Streaming API events (API version 36.0 and earlier)
\item Daily durable Streaming API events (API version 37.0 and later)
\item Streaming API concurrent clients (API version 36.0 and earlier)
\item Durable Streaming API concurrent clients (API version 37.0 and later)
\item Daily generic streaming events (API version 36.0 and earlier)
\item Daily durable generic streaming events (API version 37.0 and later)
\item Daily number of mass emails that are sent to external email addresses by using Apex or APIs
\item Daily number of single emails that are sent to external email addresses by using Apex or APIs
\item Concurrent REST API requests for results of asynchronous report runs
\item Concurrent synchronous report runs via REST API
\item Hourly asynchronous report runs via REST API
\item Hourly synchronous report runs via REST API
\item Hourly dashboard refreshes via REST API
\item Hourly REST API requests for dashboard results
\item Hourly dashboard status requests via REST API
\item Daily workflow emails
\item Hourly workflow time triggers
\item Hourly OData callouts
\item Daily and active scratch org counts
\item Data storage (MB)
\item File storage (MB)
}
}
\examples{
\dontrun{
sf_list_api_limits()
}
}
