% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-org.R
\name{sf_list_resources}
\alias{sf_list_resources}
\title{List the Resources for an API}
\usage{
sf_list_resources()
}
\value{
\code{list}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Lists available resources for the specified API version, including resource
name and URI.
}
\examples{
\dontrun{
sf_list_resources()
}
}
