% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-org.R
\name{sf_list_rest_api_versions}
\alias{sf_list_rest_api_versions}
\title{List REST API Versions}
\usage{
sf_list_rest_api_versions()
}
\value{
\code{list}
}
\description{
\lifecycle{stable}

Lists summary information about each Salesforce version currently available,
including the version, label, and a link to each version\'s root
}
\examples{
\dontrun{
sf_list_rest_api_versions()
}
}
