% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{sf_query}
\alias{sf_query}
\title{SOQL Query}
\usage{
sf_query(soql, target_object = NULL, queryall = FALSE, batch_size = 500,
  api_type = c("REST", "SOAP", "Bulk", "Async"), next_records_url = NULL)
}
\arguments{
\item{soql}{placeholder}

\item{target_object}{placeholder}

\item{queryall}{placeholder}

\item{batch_size}{placeholder}

\item{api_type}{placeholder}

\item{next_records_url}{placeholder}
}
\value{
\code{tibble}
}
\description{
Query Salesforce
}
