% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk-query.R
\name{sf_query_result_bulk_v1}
\alias{sf_query_result_bulk_v1}
\title{Retrieve the results of a Bulk 1.0 query}
\usage{
sf_query_result_bulk_v1(
  job_id,
  guess_types = TRUE,
  batch_id = NULL,
  result_id = NULL,
  api_type = c("Bulk 1.0"),
  verbose = FALSE
)
}
\arguments{
\item{job_id}{\code{character}; the Salesforce Id assigned to a submitted job as 
returned by \link{sf_create_job_bulk}. It will start with \code{"750"}.}

\item{guess_types}{\code{logical}; indicating whether or not to use \code{col_guess()} 
to try and cast the data returned in the query recordset. \code{TRUE} uses 
\code{col_guess()} and \code{FALSE} returns all values as character strings.}

\item{batch_id}{\code{character}; the Salesforce Id assigned to a submitted 
batch as returned by \link{sf_create_batches_bulk}. It will start with \code{"751"}.}

\item{result_id}{\code{character}; a string returned from 
\link{sf_batch_details_bulk} when a query has completed and specifies how to 
get the recordset}

\item{api_type}{\code{character}; one of \code{"REST"}, \code{"SOAP"}, 
\code{"Bulk 1.0"}, or \code{"Bulk 2.0"} indicating which API to use when 
making the request.}

\item{verbose}{\code{logical}; an indicator of whether to print additional 
detail for each API call, which is useful for debugging. More specifically, when 
set to \code{TRUE} the URL, header, and body will be printed for each request, 
along with additional diagnostic information where available.}
}
\value{
\code{tbl_df}, formatted by Salesforce, containing query results
}
\description{
This function returns the row-level recordset of a Bulk 1.0 query
which has already been submitted to Bulk API Job and has Completed state
}
\examples{
\dontrun{
my_query <- "SELECT Id, Name FROM Account LIMIT 1000"
job_info <- sf_create_job_bulk(operation = 'query', object = 'Account', api_type="Bulk 1.0")
query_info <- sf_submit_query_bulk(job_id = job_info$id, soql = my_query, api_type="Bulk 1.0")
result <- sf_batch_details_bulk(job_id = query_info$jobId,
                                batch_id = query_info$id)
recordset <- sf_query_result_bulk(job_id = query_info$jobId,
                                  batch_id = query_info$id,
                                  result_id = result$result)
sf_close_job_bulk(job_info$id, api_type="Bulk 1.0")
}
}
\references{
\href{https://developer.salesforce.com/docs/atlas.en-us.api_asynch.meta/api_asynch/asynch_api_bulk_query_intro.htm}{Bulk 1.0 documentation}
}
