% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-dashboard.R
\name{sf_refresh_dashboard}
\alias{sf_refresh_dashboard}
\title{Refresh an existing dashboard}
\usage{
sf_refresh_dashboard(dashboard_id, dashboard_filters = c(character(0)))
}
\arguments{
\item{dashboard_id}{\code{character}; the Salesforce Id assigned to a created
dashboard. It will start with \code{"01Z"}.}

\item{dashboard_filters}{\code{character}; Dashboard results are always
unfiltered, unless you have specified filter parameters in your request. Use
this argument to include up to three optional filter Ids. You can obtain the
list of defined filter Ids from the dashboard metadata using
\link{sf_describe_dashboard}.}
}
\value{
\code{list}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
