% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-query.R
\name{sf_reorder_cols}
\alias{sf_reorder_cols}
\title{Reorder resultset columns to prioritize \code{sObject} and \code{Id}}
\usage{
sf_reorder_cols(df)
}
\arguments{
\item{df}{\code{tbl_df}; the data frame to rearrange columns in}
}
\value{
\code{tbl_df} the formatted data frame
}
\description{
This function accepts a \code{tbl_df} with columns rearranged.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
