% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-report.R
\name{sf_report_create}
\alias{sf_report_create}
\title{Create a report}
\usage{
sf_report_create(name, report_type)
}
\arguments{
\item{name}{\code{character}; a user-specified name for the report.}

\item{report_type}{\code{character}; a character representing the type of 
report to retrieve the metadata information on. A list of valid options for 
this parameter can be gotten using \link{sf_report_types_list}.}
}
\value{
\code{list}
}
\description{
Create a report
}
