% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-report.R
\name{sf_report_delete}
\alias{sf_report_delete}
\title{Delete a report}
\usage{
sf_report_delete(report_id)
}
\arguments{
\item{report_id}{\code{character}; the Salesforce Id assigned to a created 
analytics report. It will start with \code{"00O"}.}
}
\value{
\code{logical}
}
\description{
Delete a report
}
