% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-report.R
\name{sf_report_describe}
\alias{sf_report_describe}
\title{Describe a report}
\usage{
sf_report_describe(report_id)
}
\arguments{
\item{report_id}{\code{character}; the Salesforce Id assigned to a created 
analytics report. It will start with \code{"00O"}.}
}
\value{
\code{list}
}
\description{
Describe a report
}
