% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-report.R
\name{sf_report_filter_operators_list}
\alias{sf_report_filter_operators_list}
\title{List report filter operators}
\usage{
sf_report_filter_operators_list()
}
\description{
List report filter operators
}
