% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-report-folder.R
\name{sf_report_folder_children}
\alias{sf_report_folder_children}
\title{Get the subfolders (children) of a report folder}
\usage{
sf_report_folder_children(report_folder_id, page_size = 10, page = NULL)
}
\description{
Get the subfolders (children) of a report folder
}
