% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-report-folder.R
\name{sf_report_folder_share_delete}
\alias{sf_report_folder_share_delete}
\title{Delete a report folder share}
\usage{
sf_report_folder_share_delete(report_folder_id, share_id)
}
\arguments{
\item{share_id}{\code{character}; the Salesforce Id assigned to a created 
analytics folder share (it will start with "004").}

\item{folder_id}{\code{character}; the Salesforce Id assigned to a created 
analytics folder (it will start with "00l").}
}
\value{
\code{logical}
}
\description{
Delete a report folder share
}
