% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-report-folder.R
\name{sf_report_folder_share_describe}
\alias{sf_report_folder_share_describe}
\title{Describe a report folder share}
\usage{
sf_report_folder_share_describe(report_folder_id, share_id)
}
\description{
Describe a report folder share
}
