% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-report-folder.R
\name{sf_report_folder_share_recipients}
\alias{sf_report_folder_share_recipients}
\title{Get report folder share recipients}
\usage{
sf_report_folder_share_recipients(
  report_folder_id,
  share_type = c("User", "Group", "Role"),
  search_term = "",
  limit = 100
)
}
\description{
Get report folder share recipients
}
