% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-report-folder.R
\name{sf_report_folder_share_update}
\alias{sf_report_folder_share_update}
\title{Update a report folder share}
\usage{
sf_report_folder_share_update(report_folder_id, share_id, body)
}
\description{
Update a report folder share
}
