% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-report-folder.R
\name{sf_report_folder_shares_add}
\alias{sf_report_folder_shares_add}
\title{Add shares to a report folder}
\usage{
sf_report_folder_shares_add(report_folder_id, body)
}
\description{
Creates new shares and appends them to the existing share list for the folder.
}
