% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-report-folder.R
\name{sf_report_folder_shares_list}
\alias{sf_report_folder_shares_list}
\title{List the shares in a report folder}
\usage{
sf_report_folder_shares_list(report_folder_id)
}
\description{
List the shares in a report folder
}
