% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-report-folder.R
\name{sf_report_folder_shares_update}
\alias{sf_report_folder_shares_update}
\title{Update the shares for a report folder}
\usage{
sf_report_folder_shares_update(report_folder_id, body)
}
\description{
Creates new shares to replace the existing shares in the share list for the folder.
}
