% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-report.R
\name{sf_report_instances_list}
\alias{sf_report_instances_list}
\title{List report instances}
\usage{
sf_report_instances_list(report_id)
}
\arguments{
\item{report_id}{\code{character}; the Salesforce Id assigned to a created 
analytics report. It will start with \code{"00O"}.}
}
\value{
\code{list}
}
\description{
List report instances
}
