% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-report.R
\name{sf_report_type_describe}
\alias{sf_report_type_describe}
\title{Describe a report type}
\usage{
sf_report_type_describe(report_type)
}
\arguments{
\item{report_type}{\code{character}; a character representing the type of 
report to retrieve the metadata information on. A list of valid options for 
this parameter can be gotten using \link{sf_report_types_list}.}
}
\value{
\code{list}
}
\description{
Describe a report type
}
