% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-report.R
\name{sf_reports_list}
\alias{sf_reports_list}
\title{List reports}
\usage{
sf_reports_list()
}
\description{
Displays a list of up to 200 tabular, matrix, or summary reports that you
recently viewed. To get a full list of reports by format, name, and other
fields, use a SOQL query on the Report object. The resource can also be used
to make a copy of a report.
}
