% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-org.R
\name{sf_reset_password}
\alias{sf_reset_password}
\title{Reset User Password}
\usage{
sf_reset_password(user_id, verbose = FALSE)
}
\arguments{
\item{user_id}{character; the unique Salesforce Id assigned to the User}

\item{verbose}{logical; do you want informative messages?}
}
\value{
\code{list}
}
\description{
Changes a user’s password to a temporary, system-generated value.
}
\examples{
\dontrun{
sf_reset_password(user_id = "0056A000000ZZZaaBBB")
}
}
