% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{sf_retrieve_rest}
\alias{sf_retrieve_rest}
\title{Retrieve records using REST API}
\usage{
sf_retrieve_rest(
  ids,
  fields,
  object_name,
  guess_types = TRUE,
  control,
  ...,
  verbose = FALSE
)
}
\description{
Retrieve records using REST API
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
