% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-org.R
\name{sf_server_timestamp}
\alias{sf_server_timestamp}
\title{Salesforce Server Timestamp}
\usage{
sf_server_timestamp()
}
\value{
\code{POSIXct} formatted timestamp
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Retrieves the current system timestamp from the API.
}
\examples{
\dontrun{
sf_server_timestamp()
}
}
