% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{sf_session_id}
\alias{sf_session_id}
\title{Return session_id resulting from Basic auth routine}
\usage{
sf_session_id(verbose = TRUE)
}
\arguments{
\item{verbose}{logical; do you want informative messages?}
}
\value{
character; a string of the sessionId element of the current authorized 
API session; otherwise NULL
}
\description{
Return session_id resulting from Basic auth routine
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
