% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-org.R
\name{sf_set_password}
\alias{sf_set_password}
\title{Set User Password}
\usage{
sf_set_password(user_id, password, verbose = FALSE)
}
\arguments{
\item{user_id}{character; the unique Salesforce Id assigned to the User}

\item{password}{character; a new password that you would like to set for the 
supplied user that complies to your organizations password requirements}

\item{verbose}{logical; do you want informative messages?}
}
\value{
\code{list}
}
\description{
Sets the specified user’s password to the specified value.
}
\examples{
\dontrun{
sf_set_password(user_id = "0056A000000ZZZaaBBB", password="password123")
}
}
