% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{sf_token}
\alias{sf_token}
\title{Produce Salesforce token}
\usage{
sf_token(verbose = FALSE)
}
\value{
a \code{request} object (an S3 class provided by \code{httr})
}
\description{
If token is not already available, call \code{\link{sf_auth}} to either load
from cache or initiate OAuth2.0 flow. Return the token -- not "bare" but,
rather, prepared for inclusion in downstream requests. Use
\code{access_token()} to reveal the actual access token, suitable for use
with \code{curl}.
}
\keyword{internal}
