% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachments.R
\name{sf_update_attachment_bulk_v1}
\alias{sf_update_attachment_bulk_v1}
\title{Update Attachments using Bulk 1.0 API}
\usage{
sf_update_attachment_bulk_v1(
  attachment_input_data,
  object_name = c("Attachment"),
  content_type = "ZIP_CSV",
  control,
  ...,
  verbose = FALSE
)
}
\description{
Update Attachments using Bulk 1.0 API
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
