% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachments.R
\name{sf_update_attachment_soap}
\alias{sf_update_attachment_soap}
\title{Update Attachment using SOAP API}
\usage{
sf_update_attachment_soap(
  attachment_input_data,
  object_name = c("Attachment"),
  control,
  ...,
  verbose = FALSE
)
}
\description{
Update Attachment using SOAP API
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
