% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{sf_update_bulk_v2}
\alias{sf_update_bulk_v2}
\title{Update Records using Bulk 2.0 API}
\usage{
sf_update_bulk_v2(
  input_data,
  object_name,
  guess_types = TRUE,
  control,
  ...,
  verbose = FALSE
)
}
\description{
Update Records using Bulk 2.0 API
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
