% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsert.R
\name{sf_upsert_bulk_v2}
\alias{sf_upsert_bulk_v2}
\title{Upsert Records using Bulk 2.0 API}
\usage{
sf_upsert_bulk_v2(input_data, object_name, external_id_fieldname,
  all_or_none = FALSE, ..., verbose = FALSE)
}
\description{
Upsert Records using Bulk 2.0 API
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
