% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-org.R
\name{sf_user_info}
\alias{sf_user_info}
\title{Return Current User Info}
\usage{
sf_user_info()
}
\value{
\code{list}
}
\description{
Retrieves personal information for the user associated with the current session.
}
\examples{
\dontrun{
sf_user_info()
}
}
