% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sf_write_csv}
\alias{sf_write_csv}
\title{Write a CSV file in format acceptable to Salesforce APIs}
\usage{
sf_write_csv(x, path)
}
\description{
Write a CSV file in format acceptable to Salesforce APIs
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
