% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{token_available}
\alias{token_available}
\title{Check token availability}
\usage{
token_available(verbose = FALSE)
}
\value{
\code{logical}
}
\description{
Check if a token is available in \code{\link{salesforcer}}'s internal
\code{.state} environment.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
