% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-query.R
\name{unnest_col}
\alias{unnest_col}
\title{Flatten list column}
\usage{
unnest_col(df, col)
}
\arguments{
\item{df}{\code{tbl_df}; a data frame with list column to be extracted into
multiple individual columns.}

\item{col}{\code{character}; the name of the column to unnest}
}
\value{
\code{tbl_df} parsed from the flattened list.
}
\description{
This function is a convenience function to handle a list column in a \code{tbl_df}.
The column is unnested wide while preserving the row count.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
