% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-metadata.R
\name{valid_metadata_list}
\alias{valid_metadata_list}
\title{List of Valid Data Types}
\usage{
valid_metadata_list()
}
\value{
code{list}; contains name and valid inputs for data types
}
\description{
A list of data types that are valid for the Metadata API service.
}
