% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-bulk.R
\name{validate_get_all_jobs_params}
\alias{validate_get_all_jobs_params}
\title{Validate Query Parameters When Getting List of All Bulk Jobs}
\usage{
validate_get_all_jobs_params(parameterized_search_list, type = "all")
}
\arguments{
\item{parameterized_search_list}{\code{list}; a list of search options to
locate Bulk API jobs.}

\item{type}{\code{character}; a string indicating which type of jobs to
include in the result.}
}
\value{
\code{character}; a complete URL (as a string) to send a request
to in order to retrieve queried jobs.
}
\description{
Validate Query Parameters When Getting List of All Bulk Jobs
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\seealso{
\url{https://developer.salesforce.com/docs/atlas.en-us.api_asynch.meta/api_asynch/query_get_all_jobs.htm}
}
\keyword{internal}
