% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-httr.R
\name{VERB_n}
\alias{VERB_n}
\title{Generic implementation of HTTP methods with retries and authentication}
\usage{
VERB_n(verb)
}
\arguments{
\item{verb}{string; The name of HTTP verb to execute}
}
\value{
The last response. Note that if the request doesn't succeed after
\code{times} then the request will fail and return the response.
}
\description{
Generic implementation of HTTP methods with retries and authentication
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
