% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-input-validation.R
\name{warn_w_errors_listed}
\alias{warn_w_errors_listed}
\title{List a vector of errors and provide a warning}
\usage{
warn_w_errors_listed(main_text = NULL, errors = NULL)
}
\arguments{
\item{main_text}{\code{character}; The text used to introduce the list of
errors, typically ending with a colon. For example: \code{"Consider fixing the 
following issue(s):"}.}

\item{errors}{\code{errors}; a vector of errors that will be formatted into
a bulleted list for the user to review with each error listed on a new line.}
}
\value{
\code{simpleError}
}
\description{
List a vector of errors and provide a warning
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
