% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-query.R
\name{xml_drop_and_unlist_recursively}
\alias{xml_drop_and_unlist_recursively}
\title{Recursively Drop \code{type} and \code{Id} attributes and flatten a list}
\usage{
xml_drop_and_unlist_recursively(x)
}
\arguments{
\item{x}{\code{list}; a list to be cleaned.}
}
\value{
\code{list} containing without \code{type} and \code{Id} fields that
are not requested as part of the query, but Salesforce provides.
}
\description{
This function wraps the \code{\link{xml_drop_and_unlist}} function
to recursively flatten and remove record type attributes from relationship
and nested queries.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
