% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-xml.R
\name{xml_nodeset_to_df}
\alias{xml_nodeset_to_df}
\title{xml_nodeset_to_df}
\usage{
xml_nodeset_to_df(this_node)
}
\arguments{
\item{this_node}{\code{xml_node}; to be parsed out}
}
\value{
\code{tbl_df} parsed from the supplied XML
}
\description{
A function specifically for parsing an XML node into a \code{data.frame}
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
