% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-xml.R
\name{xmlToList2}
\alias{xmlToList2}
\title{xmlToList2}
\usage{
xmlToList2(node)
}
\arguments{
\item{node}{the XML node or document to be converted to an R list}
}
\value{
\code{list} parsed from the supplied node
}
\description{
This function is an early and simple approach to converting an 
XML node or document into a more typical R list containing the data values. 
It differs from xmlToList by not including attributes at all in the output.
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
