% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-control.R
\name{accepted_controls_by_operation}
\alias{accepted_controls_by_operation}
\title{Return the Accepted Control Arguments by Operation}
\usage{
accepted_controls_by_operation(
  operation = c("create", "insert", "update", "upsert", "delete", "undelete",
    "hardDelete", "query", "queryall", "retrieve", "convertLead", "merge",
    "describeSObjects", "resetPassword")
)
}
\description{
Return the Accepted Control Arguments by Operation
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
