% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-control.R
\name{accepted_controls_by_operation}
\alias{accepted_controls_by_operation}
\title{Return the Accepted Control Arguments by Operation}
\usage{
accepted_controls_by_operation(
  operation = c("delete", "undelete", "hardDelete", "convertLead", "merge", "insert",
    "update", "upsert", "query", "queryall", "retrieve", "resetPassword",
    "describeSObjects")
)
}
\description{
Return the Accepted Control Arguments by Operation
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
