% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-httr.R
\name{catch_errors}
\alias{catch_errors}
\title{Function to catch and print HTTP errors}
\usage{
catch_errors(x)
}
\description{
Function to catch and print HTTP errors
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
