% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachments.R
\name{check_and_encode_files}
\alias{check_and_encode_files}
\title{Check that file paths exist and data is encoded if specified}
\usage{
check_and_encode_files(dat, column = "Body", encode = TRUE, n_check = 100)
}
\description{
Check that file paths exist and data is encoded if specified
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
