% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-control.R
\name{filter_valid_controls}
\alias{filter_valid_controls}
\title{Filter Out Control Arguments by API or Operation}
\usage{
filter_valid_controls(supplied, api_type = NULL, operation = NULL)
}
\description{
Filter Out Control Arguments by API or Operation
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
