% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_headers_for_verbose}
\alias{format_headers_for_verbose}
\title{Format Headers for Printing}
\usage{
format_headers_for_verbose(request_headers)
}
\arguments{
\item{request_headers}{\code{list}; a list of values from the API request or
response that represent the headers of the call}
}
\value{
\code{character}; a string constructed from the input that is easier
to read when we print it out
}
\description{
Format Headers for Printing
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
