% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_os}
\alias{get_os}
\title{Determine the host operating system}
\usage{
get_os()
}
\value{
A character string
}
\description{
This function determines whether the system running the R code
is Windows, Mac, or Linux
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\examples{
\dontrun{
get_os()
}
}
\seealso{
\url{http://conjugateprior.org/2015/06/identifying-the-os-from-r}
}
\keyword{internal}
