% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-analytics-report-folder.R
\name{make_analytics_folder_collections_url}
\alias{make_analytics_folder_collections_url}
\title{Analytics Folder collections URL generator}
\usage{
make_analytics_folder_collections_url()
}
\value{
\code{character}; a complete URL (as a string) that will be used to
send Analytics folder calls to. This URL is specific to your instance
and the API version being used.
}
\description{
Analytics Folder collections URL generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
