% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-analytics-report-folder.R
\name{make_analytics_folder_share_recipients_url}
\alias{make_analytics_folder_share_recipients_url}
\title{Analytics Folder share recipients URL generator}
\usage{
make_analytics_folder_share_recipients_url(report_folder_id, share_type)
}
\description{
Analytics Folder share recipients URL generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
