% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-bulk.R
\name{make_bulk_batch_details_url}
\alias{make_bulk_batch_details_url}
\title{Bulk Batch Details URL Generator}
\usage{
make_bulk_batch_details_url(job_id, batch_id, api_type = c("Bulk 1.0"))
}
\description{
Bulk Batch Details URL Generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
