% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-bulk.R
\name{make_bulk_batch_status_url}
\alias{make_bulk_batch_status_url}
\title{Bulk Batch Status URL Generator}
\usage{
make_bulk_batch_status_url(job_id, batch_id, api_type = c("Bulk 1.0"))
}
\arguments{
\item{job_id}{\code{character}; the Salesforce Id assigned to a submitted job as
returned by \link{sf_create_job_bulk}. It will start with \code{"750"}.}

\item{batch_id}{\code{character}; the Salesforce Id assigned to a submitted
batch as returned by \link{sf_create_batches_bulk}. It will start with \code{"751"}.}

\item{api_type}{\code{character}; a string indicating which Bulk API to execute
the call against.}
}
\value{
\code{character}; a complete URL (as a string) that will be used to
send subsequent Bulk API calls to. This URL is specific to your instance and
the API version being used.
}
\description{
Bulk Batch Status URL Generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
