% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-bulk.R
\name{make_bulk_create_job_url}
\alias{make_bulk_create_job_url}
\title{Bulk Create Job URL Generator}
\usage{
make_bulk_create_job_url(
  api_type = c("Bulk 1.0", "Bulk 2.0"),
  query_operation = FALSE
)
}
\description{
Bulk Create Job URL Generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
