% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-bulk.R
\name{make_bulk_get_all_jobs_url}
\alias{make_bulk_get_all_jobs_url}
\title{Bulk Get All Jobs Generic URL Generator}
\usage{
make_bulk_get_all_jobs_url(
  parameterized_search_list = list(isPkChunkingEnabled = NULL, jobType = NULL),
  next_records_url = NULL,
  api_type = c("Bulk 2.0")
)
}
\description{
Bulk Get All Jobs Generic URL Generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
