% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-bulk.R
\name{make_bulk_query_result_url}
\alias{make_bulk_query_result_url}
\title{Bulk Query Result URL Generator}
\usage{
make_bulk_query_result_url(
  job_id,
  batch_id = NULL,
  result_id = NULL,
  api_type = c("Bulk 1.0", "Bulk 2.0")
)
}
\arguments{
\item{job_id}{\code{character}; the Salesforce Id assigned to a submitted job as
returned by \link{sf_create_job_bulk}. It will start with \code{"750"}.}

\item{batch_id}{\code{character}; the Salesforce Id assigned to a submitted
batch as returned by \link{sf_create_batches_bulk}. It will start with \code{"751"}.}

\item{result_id}{\code{character}; the Salesforce Id assigned to a generated
result for a bulk query batch.}

\item{api_type}{\code{character}; a string indicating which Bulk API to execute
the call against.}
}
\value{
\code{character}; a complete URL (as a string) that will be used to
send subsequent Bulk API calls to. This URL is specific to your instance and
the API version being used.
}
\description{
Bulk Query Result URL Generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
