% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-bulk.R
\name{make_bulk_query_result_url}
\alias{make_bulk_query_result_url}
\title{Bulk Query Result URL Generator}
\usage{
make_bulk_query_result_url(
  job_id,
  batch_id = NULL,
  result_id = NULL,
  api_type = c("Bulk 1.0", "Bulk 2.0")
)
}
\description{
Bulk Query Result URL Generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
