% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-rest.R
\name{make_chatter_users_url}
\alias{make_chatter_users_url}
\title{Chatter Users URL Generator}
\usage{
make_chatter_users_url()
}
\value{
\code{character}; a complete URL (as a string) that will be used to
send REST API calls regarding chatter users. This URL is specific to your
instance and the API version because it relies on the base rest URL.
}
\description{
Chatter Users URL Generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
