% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-rest.R
\name{make_composite_batch_url}
\alias{make_composite_batch_url}
\title{Composite Batch URL Generator}
\usage{
make_composite_batch_url()
}
\value{
\code{character}; a complete URL (as a string) that will be used to
send composite batch REST API calls to. This URL is specific to your instance
and the API version being used because it relies on the base REST URL.
}
\description{
Composite Batch URL Generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
