% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-rest.R
\name{make_composite_url}
\alias{make_composite_url}
\title{Composite URL Generator}
\usage{
make_composite_url()
}
\description{
Composite URL Generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
