% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-analytics-dashboard.R
\name{make_dashboard_copy_url}
\alias{make_dashboard_copy_url}
\title{Dashboard Copy URL generator}
\usage{
make_dashboard_copy_url(dashboard_id)
}
\arguments{
\item{dashboard_id}{\code{character}; the Salesforce Id assigned to a created
dashboard. It will start with \code{"01Z"}.}
}
\value{
\code{character}; a complete URL (as a string) that will be used to
send Reports and Dashboards API calls to. This URL is specific to your instance
and the API version being used.
}
\description{
Dashboard Copy URL generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
