% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-analytics-dashboard.R
\name{make_dashboard_filter_operators_list_url}
\alias{make_dashboard_filter_operators_list_url}
\title{Dashboard filter operators list URL generator}
\usage{
make_dashboard_filter_operators_list_url(for_dashboards = FALSE)
}
\description{
Dashboard filter operators list URL generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
