% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-rest.R
\name{make_query_url}
\alias{make_query_url}
\title{Query URL Generator}
\usage{
make_query_url(soql, queryall, next_records_url)
}
\description{
Query URL Generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
