% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-analytics-report.R
\name{make_report_instance_url}
\alias{make_report_instance_url}
\title{Report Instance URL generator}
\usage{
make_report_instance_url(report_id, report_instance_id)
}
\description{
Report Instance URL generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
