% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-analytics-report.R
\name{make_report_type_describe_url}
\alias{make_report_type_describe_url}
\title{Report Type Describe URL generator}
\usage{
make_report_type_describe_url(type)
}
\description{
Report Type Describe URL generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
