% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-analytics-report.R
\name{make_report_types_list_url}
\alias{make_report_types_list_url}
\title{Report Type List URL generator}
\usage{
make_report_types_list_url()
}
\description{
Report Type List URL generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
