% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-analytics-report.R
\name{make_reports_list_url}
\alias{make_reports_list_url}
\title{Reports List URL generator}
\usage{
make_reports_list_url()
}
\value{
\code{character}; a complete URL (as a string) that will be used to
send Reports and Dashboards API calls to. This URL is specific to your instance
and the API version being used.
}
\description{
Reports List URL generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
