% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-rest.R
\name{make_rest_record_url}
\alias{make_rest_record_url}
\title{REST Individual Record URL Generator}
\usage{
make_rest_record_url(object, sf_id)
}
\description{
REST Individual Record URL Generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
