% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-rest.R
\name{make_search_url}
\alias{make_search_url}
\title{Search URL Generator}
\usage{
make_search_url(search_string)
}
\arguments{
\item{search_string}{\code{character}; a valid string for conducting a SOSL
search.}
}
\value{
\code{character}; a complete URL (as a string) that has applied the
proper escaping and formatting for the search specified by the string.
}
\description{
Search URL Generator
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
