% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_verbose_httr_message}
\alias{make_verbose_httr_message}
\title{Format Verbose Call}
\usage{
make_verbose_httr_message(method, url, headers = NULL, body = NULL)
}
\description{
Format Verbose Call
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
