% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{rforcecom.login}
\alias{rforcecom.login}
\title{The \code{salesforcer} backwards compatible version of
\code{RForcecom::rforcecom.login}}
\usage{
rforcecom.login(
  username,
  password,
  loginURL = "https://login.salesforce.com/",
  apiVersion = "35.0"
)
}
\arguments{
\item{username}{Your username for login to the Salesforce.com. In many cases, username is your E-mail address.}

\item{password}{Your password for login to the Salesforce.com. Note: DO NOT FORGET your Security Token. (Ex.) If your password is "Pass1234" and your security token is "XYZXYZXYZXYZ", you should set "Pass1234XYZXYZXYZXYZ".}

\item{loginURL}{(optional) Login URL. If your environment is sandbox specify (ex:) "https://test.salesforce.com".}

\item{apiVersion}{(optional) Version of the REST API and SOAP API that you want to use. (ex:) "35.0" Supported versions from v20.0 and up.}
}
\value{
\item{sessionID}{Session ID.}
\item{instanceURL}{Instance URL.}
\item{apiVersion}{API Version.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#soft-deprecated}{\figure{lifecycle-soft-deprecated.svg}{options: alt='[Soft-deprecated]'}}}{\strong{[Soft-deprecated]}}
}
