% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{rforcecom.query}
\alias{rforcecom.query}
\title{The \code{salesforcer} backwards compatible version of
\code{RForcecom::rforcecom.query}}
\usage{
rforcecom.query(session, soqlQuery, queryAll = FALSE)
}
\arguments{
\item{session}{\code{list}; a list containing \code{"sessionID"},
\code{"instanceURL"}, and \code{"apiVersion"} as returned by
\code{RForcecom::rforcecom.login}. This argument is ignored in all
backward compatible calls because the authorization credentials are stored
in an environment internal to the salesforcer package, so it is no longer
necessary to pass the session in each function call.}

\item{soqlQuery}{\code{character}; a string defining a SOQL query
(e.g. "SELECT Id, Name FROM Account").}

\item{queryAll}{\code{logical}; indicating if the query recordset should include
deleted and archived records (available only when querying Task and Event records)}
}
\value{
Result dataset.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#soft-deprecated}{\figure{lifecycle-soft-deprecated.svg}{options: alt='[Soft-deprecated]'}}}{\strong{[Soft-deprecated]}}
}
