% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{rforcecom.search}
\alias{rforcecom.search}
\title{The \code{salesforcer} backwards compatible version of
\code{\link[RForcecom]{rforcecom.search}}}
\usage{
rforcecom.search(session, queryString)
}
\arguments{
\item{session}{\code{list}; a list containing \code{"sessionID"},
\code{"instanceURL"}, and \code{"apiVersion"} as returned by
\code{\link[RForcecom]{rforcecom.login}}. This argument is ignored in all
backward compatible calls because the authorization credentials are stored
in an environment internal to the salesforcer package, so it is no longer
necessary to pass the session in each function call.}

\item{queryString}{Query strings to search. (ex: "United", "Electoronics")}
}
\description{
\lifecycle{soft-deprecated}
}
