% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{rforcecom.upsert}
\alias{rforcecom.upsert}
\title{salesforcer's backwards compatible version of rforcecom.upsert}
\usage{
rforcecom.upsert(session, objectName, externalIdField, externalId, fields)
}
\arguments{
\item{session}{\code{list}; a list containing "sessionID", "instanceURL", and "
apiVersion" as returned by \code{RForcecom::rforcecom.login()}. This argument is
ignored in all backward compatible calls because the authorization credentials 
are stored in an environment internal to the salesforcer package, so it is no longer 
necessary to pass the session in each function call.}

\item{objectName}{character; the name of the Salesforce object that the 
function is operating against (e.g. "Account", "Contact", "CustomObject__c")}

\item{externalIdField}{An external Key's field name. (ex: "AccountMaster__c")}

\item{externalId}{An external Key's ID. (ex: "999x000000xxxxxZZZ")}

\item{fields}{Field names and values. (ex: Name="CompanyName", Phone="000-000-000" )}
}
\value{
\code{NULL} if successful otherwise the function errors out
}
\description{
salesforcer's backwards compatible version of rforcecom.upsert
}
