% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salesforcer-package.R
\docType{package}
\name{salesforcer-package}
\alias{salesforcer}
\alias{salesforcer-package}
\title{\code{salesforcer} package}
\description{
An R package connecting to Salesforce APIs using tidy principles
}
\details{
A package that connects R to Salesforce via REST, SOAP, Bulk, Reports and
Dashboards, and Metadata APIs with an emphasis on the use of tidy data
principles and the tidyverse.

Additional material can be found in the
\href{https://github.com/stevenmmortimer/salesforcer}{README} on GitHub and
the package website \url{https://stevenmmortimer.github.io/salesforcer/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/StevenMMortimer/salesforcer}
  \item Report bugs at \url{https://github.com/StevenMMortimer/salesforcer/issues}
}

}
\author{
\strong{Maintainer}: Steven M. Mortimer \email{mortimer.steven.m@gmail.com}

Other contributors:
\itemize{
  \item Takekatsu Hiramura \email{thira@plavox.info} [contributor]
  \item Jennifer Bryan \email{jenny@rstudio.com} [contributor, copyright holder]
  \item Joanna Zhao \email{joanna.zhao@alumni.ubc.ca} [contributor, copyright holder]
}

}
\keyword{internal}
