% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-notification.R
\name{sf_analytics_notification_create}
\alias{sf_analytics_notification_create}
\title{Create an analytics notification}
\usage{
sf_analytics_notification_create(body)
}
\arguments{
\item{body}{\code{list}; a list of data to be formatted as JSON before being
sent to an API endpoint. This is NOT validated and is only experimental in
allowing users to utilize the API functionality before wrapper functions
have been written in R to format and validate the request in a more
user-friendly format.}
}
\value{
\code{list}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
