% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-notification.R
\name{sf_analytics_notification_delete}
\alias{sf_analytics_notification_delete}
\title{Delete an analytics notification}
\usage{
sf_analytics_notification_delete(notification_id)
}
\arguments{
\item{notification_id}{\code{character}; the Salesforce Id assigned to a created
analytics notification. It will start with \code{"0Au"}.}
}
\value{
\code{logical}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
