% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-notification.R
\name{sf_analytics_notification_describe}
\alias{sf_analytics_notification_describe}
\title{Describe an analytics notification}
\usage{
sf_analytics_notification_describe(notification_id)
}
\arguments{
\item{notification_id}{\code{character}; the Salesforce Id assigned to a created
analytics notification. It will start with \code{"0Au"}.}
}
\value{
\code{list}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
