% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-notification.R
\name{sf_analytics_notifications_limits}
\alias{sf_analytics_notifications_limits}
\title{Return limits of analytics notifications}
\usage{
sf_analytics_notifications_limits(
  source = c("lightningDashboardSubscribe", "lightningReportSubscribe",
    "waveNotification"),
  record_id = NULL
)
}
\arguments{
\item{source}{\code{character}; a vector of character to specify what type of
analytics notification to return. Valid values are:
\describe{
\item{lightningDashboardSubscribe}{Dashboard subscriptions}
\item{lightningReportSubscribe}{Report subscriptions}
\item{waveNotification}{Einstein Analytics notifications}
}
By default, all 3 sources will be returned in the results.}

\item{record_id}{\code{character}; specify if wanting to return the notification
details for a single record. Valid values are either a unique report Id
(these reference Ids start with "00O") or an Einstein Analytics lens ID (these
reference Ids start with "0FK").}
}
\value{
\code{list}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
