% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-notification.R
\name{sf_analytics_notifications_list}
\alias{sf_analytics_notifications_list}
\title{List analytics notifications}
\usage{
sf_analytics_notifications_list(
  source = c("lightningDashboardSubscribe", "lightningReportSubscribe",
    "waveNotification"),
  owner_id = NULL,
  record_id = NULL
)
}
\arguments{
\item{source}{\code{character}; a vector of character to specify what type of
analytics notification to return. Valid values are:
\describe{
\item{lightningDashboardSubscribe}{Dashboard subscriptions}
\item{lightningReportSubscribe}{Report subscriptions}
\item{waveNotification}{Einstein Analytics notifications}
}
By default, all 3 sources will be returned in the results.}

\item{owner_id}{\code{character}; Allows users with Manage Analytics Notifications
permission to get notifications for another user with the specified Salesforce
Id representing the owner, which is a User record (i.e. the Id will start with "005").}

\item{record_id}{\code{character}; specify if wanting to return the notification
details for a single record. Valid values are either a unique report Id
(these reference Ids start with "00O") or an Einstein Analytics lens ID (these
reference Ids start with "0FK").}
}
\value{
\code{list}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
