% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk-operation.R
\name{sf_bulk_create_job}
\alias{sf_bulk_create_job}
\title{Create Bulk API Job}
\usage{
sf_bulk_create_job(operation = c("insert", "delete", "upsert", "update",
  "hardDelete", "query"), object, contentType = c("CSV", "ZIP_CSV", "ZIP_XML",
  "ZIP_JSON"), externalIdFieldName = NULL, concurrencyMode = c("Parallel",
  "Serial"), lineEnding = NULL, columnDelimiter = NULL)
}
\arguments{
\item{operation}{a character string defining the type of operation being performed}

\item{object}{character; the name of the Salesforcer object that the 
function is operating against}

\item{contentType}{a character string being one of 'CSV','ZIP_CSV','ZIP_XML', or 'ZIP_JSON'}

\item{externalIdFieldName}{a character string identifying a custom field that has the "External ID" attribute on the target object. 
This field is only used when performing upserts. It will be ignored for all other operations.}

\item{concurrencyMode}{a character string either "Parallel" or "Serial" that specifies whether batches should be completed
sequentially or in parallel. Use "Serial" only if Lock contentions persist with in "Parallel" mode.}

\item{lineEnding}{a character string indicating the The line ending used for CSV job data, 
marking the end of a data row. The default is "LF", but could be "CRLF" for Windows created files.}

\item{columnDelimiter}{a character string indicating the column delimiter used for CSV job data. 
The default value is COMMA. Valid values are: "BACKQUOTE", "CARET", "COMMA", "PIPE", 
"SEMICOLON", and "TAB".}
}
\value{
A \code{list} parameters defining the created job, including id
}
\description{
This function initializes a Job in the Salesforce Bulk API
}
\examples{
\dontrun{
# insert into Account
job_info <- sf_bulk_create_job(operation='insert', object='Account')

# delete from Account
job_info <- sf_bulk_create_job(operation='delete', object='Account')

# update into Account
job_info <- sf_bulk_create_job(operation='update', object='Account')

# upsert into Account
job_info <- sf_bulk_create_job(operation='upsert',
                               externalIdFieldName='My_External_Id__c',
                               object='Account')

# insert attachments
job_info <- sf_bulk_create_job(operation='insert', object='Attachment')
}
}
\references{
\url{https://developer.salesforce.com/docs/atlas.en-us.api_asynch.meta/api_asynch/}
}
