% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk-operation.R
\name{sf_bulk_operation}
\alias{sf_bulk_operation}
\title{Run Bulk Operation}
\usage{
sf_bulk_operation(input_data, object, operation = c("insert", "delete",
  "upsert", "update"), wait_for_results = TRUE, interval_seconds = 3,
  max_attempts = 200, verbose = FALSE)
}
\arguments{
\item{input_data}{\code{named vector}, \code{matrix}, \code{data.frame}, or 
\code{tbl_df}; data can be coerced into .csv file for submitting as batch request}

\item{object}{character; the name of the Salesforcer object that the 
function is operating against}

\item{operation}{character string defining the type of operation being performed}

\item{wait_for_results}{logical; indicating whether to wait for the operation to complete 
so that the batch results of individual records can be obtained}

\item{interval_seconds}{an integer defining the seconds between attempts to check for job completion}

\item{max_attempts}{an integer defining then max number attempts to check for job completion before stopping}

\item{verbose}{logical; do you want informative messages?}
}
\description{
This function is a convenience wrapper for submitting bulk API jobs
}
\examples{
\dontrun{
n <- 20
new_contacts <- data.frame(FirstName = rep("Test", n), 
                           LastName = paste0("Contact", 1:n))
# insert new records into the Contact object
inserts <- sf_bulk_operation(input_data=new_contacts, 
                             object="Contact", 
                             operation="insert")
}
}
\references{
\url{https://developer.salesforce.com/docs/atlas.en-us.api_asynch.meta/api_asynch/}
}
