% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{sf_create}
\alias{sf_create}
\title{Create Records}
\usage{
sf_create(input_data, object, all_or_none = FALSE, api_type = c("REST",
  "SOAP", "Bulk", "Async"), verbose = FALSE)
}
\arguments{
\item{input_data}{\code{named vector}, \code{matrix}, \code{data.frame}, or 
\code{tbl_df}; data can be coerced into a \code{data.frame}}

\item{object}{character; the name of the Salesforcer object that the 
function is operating against}

\item{all_or_none}{logical; allows a call to roll back all changes unless all 
records are processed successfully.}

\item{api_type}{character; one of "REST", "SOAP", "Bulk", or "Async" indicating 
which API to use when making the request}

\item{verbose}{logical; do you want informative messages?}
}
\value{
\code{tibble}
}
\description{
Adds one or more new records to your organization’s data.
}
\examples{
\dontrun{
n <- 500
new_contacts <- tibble(FirstName = rep("Test", n), 
                       LastName = paste0("Contact", 1:n))
sf_create(new_contacts, object="Contact")                            
}
}
