% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachments.R
\name{sf_create_attachment_rest}
\alias{sf_create_attachment_rest}
\title{Create Attachment using REST API}
\usage{
sf_create_attachment_rest(
  input_data,
  object_name = "Attachment",
  control,
  ...,
  verbose = FALSE
)
}
\description{
Create Attachment using REST API
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
