% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-dashboard.R
\name{sf_dashboard_components_describe}
\alias{sf_dashboard_components_describe}
\title{Describe dashboard components}
\usage{
sf_dashboard_components_describe(dashboard_id, component_ids = c(character(0)))
}
\arguments{
\item{dashboard_id}{\code{character}; the Salesforce Id assigned to a created
dashboard. It will start with \code{"01Z"}.}

\item{component_ids}{\code{character}; a vector of Unique Salesforce Ids of a
dashboard component.}
}
\value{
\code{list}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
