% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe-metadata.R
\name{sf_describe_metadata}
\alias{sf_describe_metadata}
\title{Describe the Metadata in an Organization}
\usage{
sf_describe_metadata(verbose = FALSE)
}
\arguments{
\item{verbose}{\code{logical}; an indicator of whether to print additional
detail for each API call, which is useful for debugging. More specifically, when
set to \code{TRUE} the URL, header, and body will be printed for each request,
along with additional diagnostic information where available.}
}
\value{
A \code{tbl_df}
}
\description{
\lifecycle{experimental}

This function returns details about the organization metadata
}
\examples{
\dontrun{
# describe metadata for the organization associated with the session
metadata_info <- sf_describe_metadata()
}
}
\references{
\url{https://developer.salesforce.com/docs/atlas.en-us.api_meta.meta/api_meta/}
}
