% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk-operation.R
\name{sf_end_job_bulk}
\alias{sf_end_job_bulk}
\title{End Bulk API Job}
\usage{
sf_end_job_bulk(job_id, end_type = c("Closed", "UploadComplete",
  "Aborted"), api_type = c("Bulk 1.0", "Bulk 2.0"), verbose = FALSE)
}
\arguments{
\item{job_id}{character; the Salesforce Id assigned to a submitted job as 
returned by \link{sf_create_job_bulk}}

\item{end_type}{character; taking a value of "Closed" or "Aborted" indicating 
how the bulk job should be ended}

\item{api_type}{character; one of "REST", "SOAP", "Bulk 1.0", "Bulk 2.0", or 
"Chatter" indicating which API to use when making the request}

\item{verbose}{logical; do you want informative messages?}
}
\description{
End Bulk API Job
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
