% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datatypes.R
\name{sf_format_date}
\alias{sf_format_date}
\title{Format Dates for Create and Update operations}
\usage{
sf_format_date(x)
}
\arguments{
\item{x}{a value representing a datetime}
}
\value{
\code{character}; a date string with the time set to midnight in the
user's system timezone and then formatted in ISO8601 per the requirements of
the Salesforce APIs.
}
\description{
Format Dates for Create and Update operations
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\seealso{
\url{https://developer.salesforce.com/docs/atlas.en-us.api_asynch.meta/api_asynch/datafiles_date_format.htm}
}
\keyword{internal}
