% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datatypes.R
\name{sf_format_datetime}
\alias{sf_format_datetime}
\title{Format Datetimes for Create and Update operations}
\usage{
sf_format_datetime(x)
}
\arguments{
\item{x}{an object, potentially, representing a datetime that should be converted
to the Salesforce standard.}
}
\value{
\code{character}; an object where any values that appear to be a date
or date time are reformatted as an ISO8601 string per the requirements of the
Salesforce APIs.
}
\description{
Format Datetimes for Create and Update operations
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\seealso{
\url{https://developer.salesforce.com/docs/atlas.en-us.api_asynch.meta/api_asynch/datafiles_date_format.htm}
}
\keyword{internal}
