% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sf_format_datetime}
\alias{sf_format_datetime}
\title{Format Datetimes for Create and Update operations}
\usage{
sf_format_datetime(x)
}
\description{
Format Datetimes for Create and Update operations
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\seealso{
\url{https://developer.salesforce.com/docs/atlas.en-us.api_bulk_v2.meta/api_bulk_v2/datafiles_date_format.htm}
}
\keyword{internal}
