% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datatypes.R
\name{sf_format_time.data.frame}
\alias{sf_format_time.data.frame}
\title{Format all Date and Datetime columns in a dataset}
\usage{
\method{sf_format_time}{data.frame}(x)
}
\arguments{
\item{x}{\code{tbl_df}; a data frame object which may or may not have columns
that represent a datetime. If so, they are cast to the ISO8601 standard per
the requirements of Salesforce APIs.}
}
\value{
\code{tbl_df}; the same data frame object with datetime values formatted.
}
\description{
Format all Date and Datetime columns in a dataset
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\seealso{
\url{https://developer.salesforce.com/docs/atlas.en-us.api_asynch.meta/api_asynch/datafiles_date_format.htm}
}
\keyword{internal}
