% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-org.R
\name{sf_list_objects}
\alias{sf_list_objects}
\title{List Organization Objects and their Metadata}
\usage{
sf_list_objects()
}
\value{
\code{list}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#stable'><img src='figures/lifecycle-stable.svg' alt='Stable lifecycle'></a>}}{\strong{Stable}}

Lists the available objects and their metadata for your organization’s data.
}
\examples{
\dontrun{
sf_list_objects()
}
}
