% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-report.R
\name{sf_report_fields}
\alias{sf_report_fields}
\title{Get list of report fields}
\usage{
sf_report_fields(intersect_with = c(character(0)))
}
\arguments{
\item{intersect_with}{\code{character} a vector of unique report IDs}
}
\description{
The Report Fields resource returns report fields available for specified
reports. Use the resource to determine the best fields for use in dashboard
filters by seeing which fields different source reports have in common.
Available in API version 40.0 and later.
}
