% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-report-folder.R
\name{sf_report_folder_describe}
\alias{sf_report_folder_describe}
\title{Describe a report folder}
\usage{
sf_report_folder_describe(report_folder_id)
}
\arguments{
\item{report_folder_id}{\code{character}; the Salesforce Id assigned to a created
analytics folder (it will start with "00l").}
}
\value{
\code{list}
}
\description{
\lifecycle{experimental}
}
