% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-report-folder.R
\name{sf_report_folder_share_delete}
\alias{sf_report_folder_share_delete}
\title{Delete a report folder share}
\usage{
sf_report_folder_share_delete(report_folder_id, share_id)
}
\description{
Delete a report folder share
}
