% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-report-folder.R
\name{sf_report_folder_share_describe}
\alias{sf_report_folder_share_describe}
\title{Describe a report folder share}
\usage{
sf_report_folder_share_describe(report_folder_id, share_id)
}
\arguments{
\item{report_folder_id}{\code{character}; the Salesforce Id assigned to a created
analytics folder (it will start with "00l").}

\item{share_id}{\code{character}; the Salesforce Id assigned to a created
analytics folder share (it will start with "004").}
}
\value{
\code{list}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
