% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-report-folder.R
\name{sf_report_folder_share_update}
\alias{sf_report_folder_share_update}
\title{Update a report folder share}
\usage{
sf_report_folder_share_update(report_folder_id, share_id, body)
}
\arguments{
\item{report_folder_id}{\code{character}; the Salesforce Id assigned to a created
analytics folder (it will start with "00l").}

\item{share_id}{\code{character}; the Salesforce Id assigned to a created
analytics folder share (it will start with "004").}

\item{body}{\code{list}; a list of data to be formatted as JSON before being
sent to an API endpoint. This is NOT validated and is only experimental in
allowing users to utilize the API functionality before wrapper functions
have been written in R to format and validate the request in a more
user-friendly format.}
}
\value{
\code{list}
}
\description{
\lifecycle{experimental}
}
