% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-report-folder.R
\name{sf_report_folder_shares_add}
\alias{sf_report_folder_shares_add}
\title{Add shares to a report folder}
\usage{
sf_report_folder_shares_add(report_folder_id, body)
}
\arguments{
\item{report_folder_id}{\code{character}; the Salesforce Id assigned to a created
analytics folder (it will start with "00l").}

\item{body}{\code{list}; a list of data to be formatted as JSON before being
sent to an API endpoint. This is NOT validated and is only experimental in
allowing users to utilize the API functionality before wrapper functions
have been written in R to format and validate the request in a more
user-friendly format.}
}
\value{
\code{list}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

Creates new shares and appends them to the existing share list for the folder.
}
