% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-report-folder.R
\name{sf_report_folder_shares_list}
\alias{sf_report_folder_shares_list}
\title{List the shares in a report folder}
\usage{
sf_report_folder_shares_list(report_folder_id)
}
\arguments{
\item{report_folder_id}{\code{character}; the Salesforce Id assigned to a created
analytics folder (it will start with "00l").}
}
\value{
\code{list}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
