% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-report.R
\name{sf_report_instance_delete}
\alias{sf_report_instance_delete}
\title{Delete a report instance}
\usage{
sf_report_instance_delete(report_id, report_instance_id, verbose = FALSE)
}
\arguments{
\item{report_id}{\code{character}; the Salesforce Id assigned to a created 
analytics report. It will start with \code{"00O"}.}

\item{report_instance_id}{\code{character}; the Salesforce Id assigned to a 
created analytics report instance (an asynchronous run). It will start with 
\code{"0LG"}.}

\item{verbose}{\code{logical}; an indicator of whether to print additional 
detail for each API call, which is useful for debugging. More specifically, when 
set to \code{TRUE} the URL, header, and body will be printed for each request, 
along with additional diagnostic information where available.}
}
\value{
\code{logical} indicating whether the report instance was deleted. This function 
will return \code{TRUE} if successful in deleting the report instance.
}
\description{
If the given report instance has a status of \code{Success} or \code{Error}, 
delete the report instance.
}
\examples{
\dontrun{
# first, get the Id of a report in your Org
all_reports <- sf_query("SELECT Id, Name FROM Report")
this_report_id <- all_reports$Id[1]

# second, ensure that report has been executed at least once asynchronously
results <- sf_report_execute(this_report_id, async=TRUE)

# check if that report has succeeded, if so (or if it errored), then delete
instance_list <- sf_report_instances_list(this_report_id)
instance_status <- instance_list[[which(instance_list$id == results$id), "status"]]
}
}
\seealso{
\href{https://developer.salesforce.com/docs/atlas.en-us.api_analytics.meta/api_analytics/sforce_analytics_rest_api_instance_resource_results.htm}{Salesforce Documentation}
}
