% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-report.R
\name{sf_report_query}
\alias{sf_report_query}
\title{Query a report}
\usage{
sf_report_query()
}
\description{
This function allows for pulling specific data from a report. There is a convenience 
function (\link{sf_get_report_data}) to get the report data in a tabular 
format returned as a \code{tbl_df}.
}
