% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-org.R
\name{sf_server_timestamp}
\alias{sf_server_timestamp}
\title{Salesforce Server Timestamp}
\usage{
sf_server_timestamp()
}
\value{
\code{POSIXct} formatted timestamp
}
\description{
Retrieves the current system timestamp from the API.
}
\examples{
\dontrun{
sf_server_timestamp()
}
}
