% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-org.R
\name{sf_user_info}
\alias{sf_user_info}
\title{Return Current User Info}
\usage{
sf_user_info(api_type = c("SOAP", "Chatter"), verbose = FALSE)
}
\arguments{
\item{api_type}{character; one of "REST", "SOAP", "Bulk 1.0", "Bulk 2.0", or 
"Chatter" indicating which API to use when making the request}

\item{verbose}{logical; do you want informative messages?}
}
\value{
\code{list}
}
\description{
Retrieves personal information for the user associated with the current session.
}
\examples{
\dontrun{
sf_user_info()
}
}
