% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-org.R
\name{sf_user_info}
\alias{sf_user_info}
\title{Return Current User Info}
\usage{
sf_user_info(api_type = c("SOAP", "Chatter"), verbose = FALSE)
}
\arguments{
\item{api_type}{\code{character}; one of \code{"REST"}, \code{"SOAP"},
\code{"Bulk 1.0"}, or \code{"Bulk 2.0"} indicating which API to use when
making the request.}

\item{verbose}{\code{logical}; an indicator of whether to print additional
detail for each API call, which is useful for debugging. More specifically, when
set to \code{TRUE} the URL, header, and body will be printed for each request,
along with additional diagnostic information where available.}
}
\value{
\code{list}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#maturing'><img src='figures/lifecycle-maturing.svg' alt='Maturing lifecycle'></a>}}{\strong{Maturing}}

Retrieves personal information for the user associated with the current session.
}
\examples{
\dontrun{
sf_user_info()
}
}
