% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sf_write_csv}
\alias{sf_write_csv}
\title{Write a CSV file in format acceptable to Salesforce APIs}
\usage{
sf_write_csv(x, file, path = deprecated())
}
\arguments{
\item{x}{\code{tbl_df}; a data frame object to save as a CSV}

\item{file}{A file or connection to write to.}

\item{path}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
use the \code{file} argument instead.}
}
\value{
the input \code{x} invisibly. This function is called for its
side-effect of creating a CSV file at the specified location using the format
required by Salesforce.
}
\description{
Write a CSV file in format acceptable to Salesforce APIs
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\keyword{internal}
