% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-bulk.R
\name{validate_get_all_jobs_params}
\alias{validate_get_all_jobs_params}
\title{Validate Query Parameters When Getting List of All Bulk Jobs}
\usage{
validate_get_all_jobs_params(parameterized_search_list, type = "all")
}
\description{
Validate Query Parameters When Getting List of All Bulk Jobs
}
\note{
This function is meant to be used internally. Only use when debugging.
}
\seealso{
\url{https://developer.salesforce.com/docs/atlas.en-us.api_bulk_v2.meta/api_bulk_v2/query_get_all_jobs.htm}
}
\keyword{internal}
